@@ -145,7 +145,7 @@ struct _OgreLodExport LodConfig {
         /// Whether you want to process it immediatelly on main thread or you want to use Ogre::WorkQueue.
         /// If you use workqueue the generator will return immediately. After processed in background,
         /// the LodWorkQueueInjector will inject it in frameEnd event when rendering next frame.
-        /// Ready LODs can also be injected by calling Root::getSingleton().getWorkQueue()->processResponses().
+        /// Ready LODs can also be injected by calling Root::getSingleton().getWorkQueue()->processMainThreadTasks().
         /// (disabled by default)
         bool useBackgroundQueue;
         /// If enabled, it allows up to 50% smaller index buffers by storing once shared faces with frame shifting.
