@@ -52,8 +52,7 @@ namespace Ogre {
     //---------------------------------------------------------------------
     OverlayManager::OverlayManager() 
       : mLastViewportWidth(0), 
-        mLastViewportHeight(0), 
-        mLastViewportOrientationMode(OR_DEGREE_0),
+        mLastViewportHeight(0),
         mPixelRatio(1)
     {
 
@@ -210,17 +209,10 @@ namespace Ogre {
     void OverlayManager::_queueOverlaysForRendering(Camera* cam, 
         RenderQueue* pQueue, Viewport* vp)
     {
-        bool orientationModeChanged = false;
-#if OGRE_NO_VIEWPORT_ORIENTATIONMODE == 0
-        orientationModeChanged = (mLastViewportOrientationMode != vp->getOrientationMode());
-#endif
         // Flag for update pixel-based GUIElements if viewport has changed dimensions
         if (mLastViewportWidth != int(vp->getActualWidth() / mPixelRatio) ||
-            mLastViewportHeight != int(vp->getActualHeight() / mPixelRatio) || orientationModeChanged)
+            mLastViewportHeight != int(vp->getActualHeight() / mPixelRatio))
         {
-#if OGRE_NO_VIEWPORT_ORIENTATIONMODE == 0
-            mLastViewportOrientationMode = vp->getOrientationMode();
-#endif
             mLastViewportWidth = int(vp->getActualWidth() / mPixelRatio);
             mLastViewportHeight = int(vp->getActualHeight() / mPixelRatio);
         }
@@ -230,13 +222,6 @@ namespace Ogre {
         for (i = mOverlayMap.begin(); i != iend; ++i)
         {
             Overlay* o = i->second;
-#if OGRE_NO_VIEWPORT_ORIENTATIONMODE == 0
-            if (orientationModeChanged)
-            {
-                // trick to trigger transform update of the overlay
-                o->scroll(0.f, 0.f);
-            }
-#endif
             o->_findVisibleObjects(cam, pQueue, vp);
         }
     }
@@ -254,15 +239,6 @@ namespace Ogre {
     Real OverlayManager::getViewportAspectRatio(void) const
     {
         return (Real)mLastViewportWidth / (Real)mLastViewportHeight;
-    }
-    //---------------------------------------------------------------------
-    OrientationMode OverlayManager::getViewportOrientationMode(void) const
-    {
-#if OGRE_NO_VIEWPORT_ORIENTATIONMODE != 0
-        OGRE_EXCEPT(Exception::ERR_NOT_IMPLEMENTED,
-                    "Getting ViewPort orientation mode is not supported");
-#endif
-        return mLastViewportOrientationMode;
     }
     //---------------------------------------------------------------------
     float OverlayManager::getPixelRatio(void) const
