@@ -49,7 +49,7 @@ namespace Ogre
         This class is used for managing terrain LOD data's loading, unloading.
     */
 
-    class _OgreTerrainExport TerrainLodManager : public WorkQueue::RequestHandler, public WorkQueue::ResponseHandler, public TerrainAlloc
+    class _OgreTerrainExport TerrainLodManager
     {
     public:
         static const uint32 TERRAINLODDATA_CHUNK_ID;
@@ -89,12 +89,6 @@ namespace Ogre
         void close();
         bool isOpen() const;
 
-        static const uint16 WORKQUEUE_LOAD_LOD_DATA_REQUEST;
-        bool canHandleRequest(const WorkQueue::Request* req, const WorkQueue* srcQ) override;
-        bool canHandleResponse(const WorkQueue::Response* res, const WorkQueue* srcQ) override;
-        WorkQueue::Response* handleRequest(const WorkQueue::Request* req, const WorkQueue* srcQ) override;
-        void handleResponse(const WorkQueue::Response* res, const WorkQueue* srcQ) override;
-
         void updateToLodLevel(int lodLevel, bool synchronous = false);
         /// Save each LOD level separately compressed so seek is possible
         static void saveLodData(StreamSerialiser& stream, Terrain* terrain);
@@ -126,6 +120,9 @@ namespace Ogre
             return mLodInfoTable[lodLevel];
         }
     private:
+        WorkQueue::Response* handleRequest(const WorkQueue::Request* req, const WorkQueue* srcQ);
+        void handleResponse(const WorkQueue::Response* res, const WorkQueue* srcQ);
+
         void init();
         void buildLodInfoTable();
 
@@ -151,7 +148,6 @@ namespace Ogre
         Terrain* mTerrain;
         DataStreamPtr mDataStream;
         size_t mStreamOffset;
-        uint16 mWorkQueueChannel;
 
         LodInfo* mLodInfoTable;
         int mTargetLodLevel;    /// Which LOD level is demanded
