@@ -37,7 +37,6 @@ THE SOFTWARE.
 
 namespace Ogre
 {
-    const uint16 TerrainGroup::WORKQUEUE_LOAD_REQUEST = 1;
     const uint32 TerrainGroup::CHUNK_ID = StreamSerialiser::makeIdentifier("TERG");
     const uint16 TerrainGroup::CHUNK_VERSION = 1;
 
@@ -60,12 +59,6 @@ namespace Ogre
         // by default we delete input data because we copy it, unless user
         // passes us an ImportData where they explicitly don't want it copied
         mDefaultImportData.deleteInputData = true;
-
-        WorkQueue* wq = Root::getSingleton().getWorkQueue();
-        mWorkQueueChannel = wq->getChannel("Ogre/TerrainGroup");
-        wq->addRequestHandler(mWorkQueueChannel, this);
-        wq->addResponseHandler(mWorkQueueChannel, this);
-
     }
     //---------------------------------------------------------------------
     TerrainGroup::TerrainGroup(SceneManager* sm)
@@ -85,11 +78,6 @@ namespace Ogre
         // by default we delete input data because we copy it, unless user
         // passes us an ImportData where they explicitly don't want it copied
         mDefaultImportData.deleteInputData = true;
-
-        WorkQueue* wq = Root::getSingleton().getWorkQueue();
-        mWorkQueueChannel = wq->getChannel("Ogre/TerrainGroup");
-        wq->addRequestHandler(mWorkQueueChannel, this);
-        wq->addResponseHandler(mWorkQueueChannel, this);
     }
     //---------------------------------------------------------------------
     TerrainGroup::~TerrainGroup()
@@ -104,15 +92,10 @@ namespace Ogre
         while (getNumTerrainPrepareRequests() > 0)
         {
             OGRE_THREAD_SLEEP(50);
-            Root::getSingleton().getWorkQueue()->processResponses();
+            Root::getSingleton().getWorkQueue()->processMainThreadTasks();
         }
 
         removeAllTerrains();
-
-        WorkQueue* wq = Root::getSingleton().getWorkQueue();
-        wq->removeRequestHandler(mWorkQueueChannel, this);
-        wq->removeResponseHandler(mWorkQueueChannel, this);
-
     }
     //---------------------------------------------------------------------
     void TerrainGroup::setOrigin(const Vector3& pos)
@@ -373,17 +356,31 @@ namespace Ogre
             // Use shared pool of buffers
             slot->instance->setGpuBufferAllocator(&mBufferAllocator);
 
-            LoadRequest req;
-            req.slot = slot;
-            req.origin = this;
             std::pair<TerrainPrepareRequestMap::iterator, bool> ret = mTerrainPrepareRequests.emplace(slot, 0);
             assert(ret.second == true);
-            WorkQueue::RequestID id =
-                Root::getSingleton().getWorkQueue()->addRequest(
-                    mWorkQueueChannel, WORKQUEUE_LOAD_REQUEST,
-                    req, 0, synchronous);
-            if (!synchronous)
-                ret.first->second = id;
+
+            if(synchronous)
+            {
+                auto r = new WorkQueue::Request(0, 0, slot, 0, 0);
+                auto res = handleRequest(r, NULL);
+                handleResponse(res, NULL);
+                delete res;
+                return;
+            }
+
+            Root::getSingleton().getWorkQueue()->addTask(
+                [this, slot]()
+                {
+                    auto r = new WorkQueue::Request(0, 0, slot, 0, 0);
+                    auto res = handleRequest(r, NULL);
+                    Root::getSingleton().getWorkQueue()->addMainThreadTask(
+                        [this, res]()
+                        {
+                            handleResponse(res, NULL);
+                            delete res;
+                        });
+                });
+            ret.first->second = 0;
         }
     }
     //---------------------------------------------------------------------
@@ -710,23 +707,12 @@ namespace Ogre
         return false;
     }
     //---------------------------------------------------------------------
-    bool TerrainGroup::canHandleRequest(const WorkQueue::Request* req, const WorkQueue* srcQ)
-    {
-        LoadRequest lreq = any_cast<LoadRequest>(req->getData());
-        // only deal with own requests
-        if (lreq.origin != this)
-            return false;
-        else
-            return RequestHandler::canHandleRequest(req, srcQ);
-
-    }
-    //---------------------------------------------------------------------
     WorkQueue::Response* TerrainGroup::handleRequest(const WorkQueue::Request* req, const WorkQueue* srcQ)
     {
-        LoadRequest lreq = any_cast<LoadRequest>(req->getData());
+        auto slot = any_cast<TerrainSlot*>(req->getData());
 
-        TerrainSlotDefinition& def = lreq.slot->def;
-        Terrain* t = lreq.slot->instance;
+        TerrainSlotDefinition& def = slot->def;
+        Terrain* t = slot->instance;
         assert(t && "Terrain instance should have been constructed in the main thread");
         WorkQueue::Response* response = 0;
         try
@@ -754,32 +740,17 @@ namespace Ogre
 
     }
     //---------------------------------------------------------------------
-    bool TerrainGroup::canHandleResponse(const WorkQueue::Response* res, const WorkQueue* srcQ)
-    {
-        LoadRequest lreq = any_cast<LoadRequest>(res->getRequest()->getData());
-        // only deal with own requests
-        if (lreq.origin != this)
-            return false;
-        else
-            return true;
-
-    }
-    //---------------------------------------------------------------------
     void TerrainGroup::handleResponse(const WorkQueue::Response* res, const WorkQueue* srcQ)
     {
-        // Data was already deleted so nothing we can do anymore.
-        if (res->getRequest()->getAborted())
-            return;
-
         // No response data, just request
-        LoadRequest lreq = any_cast<LoadRequest>(res->getRequest()->getData());
+        auto slot = any_cast<TerrainSlot*>(res->getRequest()->getData());
 
-        TerrainPrepareRequestMap::iterator it = mTerrainPrepareRequests.find(lreq.slot);
+        TerrainPrepareRequestMap::iterator it = mTerrainPrepareRequests.find(slot);
 
         // This slot was scheduled to be deleted.
         if (it == mTerrainPrepareRequests.end())
         {
-            freeTerrainSlotInstance(lreq.slot);
+            freeTerrainSlotInstance(slot);
             return;
         }
         else
@@ -789,7 +760,6 @@ namespace Ogre
 
         if (res->succeeded())
         {
-            TerrainSlot* slot = lreq.slot;
             Terrain* terrain = slot->instance;
             if (terrain)
             {
@@ -819,9 +789,9 @@ namespace Ogre
         {
             // oh dear
             LogManager::getSingleton().stream(LML_CRITICAL) <<
-                "We failed to prepare the terrain at (" << lreq.slot->x << ", " <<
-                lreq.slot->y <<") with the error '" << res->getMessages() << "'";
-            freeTerrainSlotInstance(lreq.slot);
+                "We failed to prepare the terrain at (" << slot->x << ", " <<
+                slot->y <<") with the error '" << res->getMessages() << "'";
+            freeTerrainSlotInstance(slot);
         }
 
     }
@@ -920,12 +890,7 @@ namespace Ogre
         // Terrain was in load request so we need to schedule the deletion see handleResponse().
         if (it != mTerrainPrepareRequests.end())
         {
-            WorkQueue::RequestID id = it->second;
             mTerrainPrepareRequests.erase(it);
-
-            // We can free immediately since this slot was aborted before it could've been processed.
-            if (Root::getSingleton().getWorkQueue()->abortPendingRequest(id))
-                slot->freeInstance();
         }
         else
         {
