@@ -621,10 +621,13 @@ namespace Ogre
             numVerts += mVertexDataRecord->size * mVertexDataRecord->numSkirtRowsCols;
             numVerts += mVertexDataRecord->size * mVertexDataRecord->numSkirtRowsCols;
             // manually create CPU-side buffer
-            HardwareVertexBufferSharedPtr posbuf(
-                OGRE_NEW DefaultHardwareVertexBuffer(dcl->getVertexSize(POSITION_BUFFER), numVerts, HardwareBuffer::HBU_STATIC_WRITE_ONLY));
-            HardwareVertexBufferSharedPtr deltabuf(
-                OGRE_NEW DefaultHardwareVertexBuffer(dcl->getVertexSize(DELTA_BUFFER), numVerts, HardwareBuffer::HBU_STATIC_WRITE_ONLY));
+            auto pos_sz = dcl->getVertexSize(POSITION_BUFFER);
+            auto posbuf = std::make_shared<HardwareVertexBuffer>(nullptr, pos_sz, numVerts,
+                                                                 new DefaultHardwareBuffer(pos_sz * numVerts));
+
+            auto delta_sz = dcl->getVertexSize(DELTA_BUFFER);
+            auto deltabuf = std::make_shared<HardwareVertexBuffer>(nullptr, delta_sz, numVerts,
+                                                                   new DefaultHardwareBuffer(delta_sz * numVerts));
 
             mVertexDataRecord->cpuVertexData->vertexStart = 0;
             mVertexDataRecord->cpuVertexData->vertexCount = numVerts;
@@ -637,8 +640,8 @@ namespace Ogre
         }
     }
     //----------------------------------------------------------------------
-    void TerrainQuadTreeNode::updateVertexBuffer(HardwareVertexBufferSharedPtr& posbuf, 
-        HardwareVertexBufferSharedPtr& deltabuf, const Rect& rect)
+    void TerrainQuadTreeNode::updateVertexBuffer(const HardwareVertexBufferPtr& posbuf,
+        const HardwareVertexBufferPtr& deltabuf, const Rect& rect)
     {
         assert (rect.left >= mOffsetX && rect.right <= mBoundaryX && 
             rect.top >= mOffsetY && rect.bottom <= mBoundaryY);
