@@ -32,7 +32,6 @@ THE SOFTWARE.
 #include "OgreVolumeIsoSurfaceMC.h"
 #include "OgreVolumeOctreeNodeSplitPolicy.h"
 #include "OgreVolumeTextureSource.h"
-#include "OgreVolumeChunkHandler.h"
 #include "OgreVolumeMeshBuilder.h"
 #include "OgreVolumeDualGridGenerator.h"
 #include "OgreSceneNode.h"
@@ -44,12 +43,12 @@ THE SOFTWARE.
 #include "OgreVolumeMeshBuilder.h"
 #include "OgreVolumeOctreeNode.h"
 #include "OgreMaterialManager.h"
+#include "OgreWorkQueue.h"
 
 namespace Ogre {
 namespace Volume {
 
     const String Chunk::MOVABLE_TYPE_NAME = "VolumeChunk";
-    ChunkHandler Chunk::mChunkHandler;
     
     //-----------------------------------------------------------------------
 
@@ -72,12 +71,19 @@ namespace Volume {
             req.maxLevels = maxLevels;
             req.isUpdate = mShared->parameters->updateFrom != Vector3::ZERO || mShared->parameters->updateTo != Vector3::ZERO;
 
-            req.origin = this;
             req.root = OGRE_NEW OctreeNode(from, to);
             req.meshBuilder = OGRE_NEW MeshBuilder();
             req.dualGridGenerator = OGRE_NEW DualGridGenerator();
 
-            mChunkHandler.addRequest(req);
+            Root::getSingleton().getWorkQueue()->addTask([this, req]() {
+                prepareGeometry(req.level, req.root, req.dualGridGenerator, req.meshBuilder, req.totalFrom, req.totalTo);
+                Root::getSingleton().getWorkQueue()->addMainThreadTask([this, req]() {
+                    loadGeometry(req.meshBuilder, req.dualGridGenerator, req.root, req.level, req.isUpdate);
+                    delete req.root;
+                    delete req.meshBuilder;
+                    delete req.dualGridGenerator;
+                });
+            });
         }
         else
         {
@@ -339,7 +345,7 @@ namespace Volume {
             while(mShared->chunksBeingProcessed)
             {
                 OGRE_THREAD_SLEEP(0);
-                mChunkHandler.processWorkQueue();
+                Root::getSingleton().getWorkQueue()->processMainThreadTasks();
             }
         }
         
