@@ -46,7 +46,7 @@ namespace Ogre {
         it possible to apply postfilter effects, HDRI postprocessing, and shadow 
         effects to a Viewport.
      */
-    class _OgreExport Compositor: public Resource
+    class _OgreExport Compositor final : public Resource
     {
     public:
         Compositor(ResourceManager* creator, const String& name, ResourceHandle handle,
