@@ -23,10 +23,10 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE
 -------------------------------------------------------------------------*/
-#ifndef __OgreDefaultWorkQueueStandard_H__
-#define __OgreDefaultWorkQueueStandard_H__
+#ifndef __OgreDefaultWorkQueue_H__
+#define __OgreDefaultWorkQueue_H__
 
-#include "../OgreWorkQueue.h"
+#include "OgreWorkQueue.h"
 
 namespace Ogre
 {
