@@ -215,10 +215,10 @@ namespace Ogre
                 declareElement(VET_FLOAT3, VES_POSITION);
             }
 
-            mTempVertex.position = pos;
+            mTempVertex.position = Vector3f(pos);
 
             // update bounds
-            mAABB.merge(mTempVertex.position);
+            mAABB.merge(pos);
             mRadius = std::max(mRadius, mTempVertex.position.length());
 
             // reset current texture coord
@@ -241,7 +241,7 @@ namespace Ogre
             {
                 declareElement(VET_FLOAT3, VES_NORMAL);
             }
-            mTempVertex.normal = norm;
+            mTempVertex.normal = Vector3f(norm);
         }
         /// @overload
         void normal(float x, float y, float z)  { normal({x, y, z}); }
@@ -260,7 +260,7 @@ namespace Ogre
             {
                 declareElement(VET_FLOAT3, VES_TANGENT);
             }
-            mTempVertex.tangent = tan;
+            mTempVertex.tangent = Vector3f(tan);
         }
 
         /// @overload
@@ -281,8 +281,7 @@ namespace Ogre
             {
                 declareElement(VET_FLOAT1, VES_TEXTURE_COORDINATES);
             }
-            mTempVertex.texCoordDims[mTexCoordIndex] = 1;
-            mTempVertex.texCoord[mTexCoordIndex].x = u;
+            mTempVertex.texCoord[mTexCoordIndex][0] = u;
 
             ++mTexCoordIndex;
         }
@@ -294,9 +293,8 @@ namespace Ogre
             {
                 declareElement(VET_FLOAT2, VES_TEXTURE_COORDINATES);
             }
-            mTempVertex.texCoordDims[mTexCoordIndex] = 2;
-            mTempVertex.texCoord[mTexCoordIndex].x = u;
-            mTempVertex.texCoord[mTexCoordIndex].y = v;
+            mTempVertex.texCoord[mTexCoordIndex][0] = u;
+            mTempVertex.texCoord[mTexCoordIndex][1] = v;
 
             ++mTexCoordIndex;
         }
@@ -308,10 +306,9 @@ namespace Ogre
             {
                 declareElement(VET_FLOAT3, VES_TEXTURE_COORDINATES);
             }
-            mTempVertex.texCoordDims[mTexCoordIndex] = 3;
-            mTempVertex.texCoord[mTexCoordIndex].x = u;
-            mTempVertex.texCoord[mTexCoordIndex].y = v;
-            mTempVertex.texCoord[mTexCoordIndex].z = w;
+            mTempVertex.texCoord[mTexCoordIndex][0] = u;
+            mTempVertex.texCoord[mTexCoordIndex][1] = v;
+            mTempVertex.texCoord[mTexCoordIndex][2] = w;
 
             ++mTexCoordIndex;
         }
@@ -329,8 +326,7 @@ namespace Ogre
             {
                 declareElement(VET_FLOAT4, VES_TEXTURE_COORDINATES);
             }
-            mTempVertex.texCoordDims[mTexCoordIndex] = 4;
-            mTempVertex.texCoord[mTexCoordIndex] = xyzw;
+            mTempVertex.texCoord[mTexCoordIndex] = Vector4f(xyzw);
 
             ++mTexCoordIndex;
         }
@@ -634,11 +630,10 @@ namespace Ogre
         /// Temporary vertex structure
         struct TempVertex
         {
-            Vector3 position;
-            Vector3 normal;
-            Vector3 tangent;
-            Vector4 texCoord[OGRE_MAX_TEXTURE_COORD_SETS];
-            ushort texCoordDims[OGRE_MAX_TEXTURE_COORD_SETS];
+            Vector3f position;
+            Vector3f normal;
+            Vector3f tangent;
+            Vector4f texCoord[OGRE_MAX_TEXTURE_COORD_SETS];
             ColourValue colour;
         };
         /// Temp storage
@@ -682,14 +677,14 @@ namespace Ogre
 
 
         /// Delete temp buffers and reset init counts
-        virtual void resetTempAreas(void);
+        void resetTempAreas(void);
         /// Resize the temp vertex buffer?
-        virtual void resizeTempVertexBufferIfNeeded(size_t numVerts);
+        void resizeTempVertexBufferIfNeeded(size_t numVerts);
         /// Resize the temp index buffer?
-        virtual void resizeTempIndexBufferIfNeeded(size_t numInds);
+        void resizeTempIndexBufferIfNeeded(size_t numInds);
 
         /// Copy current temp vertex into buffer
-        virtual void copyTempVertexToBuffer(void);
+        void copyTempVertexToBuffer(void);
 
     private:
         void declareElement(VertexElementType t, VertexElementSemantic s);
