@@ -138,7 +138,7 @@ namespace Ogre {
         _keyFrameListChanged();
     }
     //---------------------------------------------------------------------
-    NumericAnimationTrack* Animation::createNumericTrack(unsigned short handle)
+    NumericAnimationTrack* Animation::createNumericTrack(unsigned short handle, const AnimableValuePtr& anim)
     {
         if (hasNumericTrack(handle))
         {
@@ -148,19 +148,9 @@ namespace Ogre {
                 "Animation::createNumericTrack");
         }
 
-        NumericAnimationTrack* ret = OGRE_NEW NumericAnimationTrack(this, handle);
+        NumericAnimationTrack* ret = OGRE_NEW NumericAnimationTrack(this, handle, anim);
 
         mNumericTrackList[handle] = ret;
-        return ret;
-    }
-    //---------------------------------------------------------------------
-    NumericAnimationTrack* Animation::createNumericTrack(unsigned short handle, 
-        const AnimableValuePtr& anim)
-    {
-        NumericAnimationTrack* ret = createNumericTrack(handle);
-
-        ret->setAssociatedAnimable(anim);
-
         return ret;
     }
     //---------------------------------------------------------------------
