@@ -52,8 +52,6 @@ namespace Ogre
 
         mShuttingDown = false;
 
-        mWorkerFunc = OGRE_NEW_T(WorkerFunc(this), MEMCATEGORY_GENERAL);
-
         LogManager::getSingleton().stream() <<
             "DefaultWorkQueue('" << mName << "') initialising on thread " <<
             OGRE_THREAD_CURRENT_ID
@@ -66,7 +64,7 @@ namespace Ogre
         mNumThreadsRegisteredWithRS = 0;
         for (size_t i = 0; i < mWorkerThreadCount; ++i)
         {
-            OGRE_THREAD_CREATE(t, *mWorkerFunc);
+            OGRE_THREAD_CREATE(t, [this]() { _threadMain(); });
             mWorkers.push_back(t);
         }
 
@@ -107,7 +105,7 @@ namespace Ogre
             << ".";
 
         mShuttingDown = true;
-        abortAllRequests();
+
 #if OGRE_THREAD_SUPPORT
         // wake all threads (they should check shutting down as first thing after wait)
         OGRE_THREAD_NOTIFY_ALL(mRequestCondition);
@@ -120,9 +118,6 @@ namespace Ogre
         }
         mWorkers.clear();
 #endif
-
-        OGRE_DELETE_T(mWorkerFunc, WorkerFunc, MEMCATEGORY_GENERAL);
-        mWorkerFunc = 0;
 
         mIsRunning = false;
     }
@@ -139,7 +134,7 @@ namespace Ogre
 #if OGRE_THREAD_SUPPORT
         // Lock; note that OGRE_THREAD_WAIT will free the lock
             OGRE_WQ_LOCK_MUTEX_NAMED(mRequestMutex, queueLock);
-        if (mRequestQueue.empty())
+        if (mTasks.empty())
         {
             // frees lock and suspends the thread
             OGRE_THREAD_WAIT(mRequestCondition, mRequestMutex, queueLock);
