@@ -726,7 +726,7 @@ namespace Ogre {
             if (!sub->getSubMesh()->useSharedVertices
                 && sub->getSubMesh()->getVertexAnimationType() != VAT_NONE)
             {
-                ret = ret && sub->_getVertexAnimTempBufferInfo()->buffersCheckedOut(
+                ret = ret && sub->mTempVertexAnimInfo.buffersCheckedOut(
                     true, sub->getSubMesh()->getVertexAnimationIncludesNormals());
             }
         }
@@ -1305,16 +1305,6 @@ namespace Ogre {
     {
         assert (mHardwareVertexAnimVertexData && "Not vertex animated or has no shared vertex data!");
         return mHardwareVertexAnimVertexData.get();
-    }
-    //-----------------------------------------------------------------------
-    TempBlendedBufferInfo* Entity::_getSkelAnimTempBufferInfo(void)
-    {
-        return &mTempSkelAnimInfo;
-    }
-    //-----------------------------------------------------------------------
-    TempBlendedBufferInfo* Entity::_getVertexAnimTempBufferInfo(void)
-    {
-        return &mTempVertexAnimInfo;
     }
     //-----------------------------------------------------------------------
     bool Entity::cacheBoneMatrices(void)
@@ -2284,5 +2274,125 @@ namespace Ogre {
 
         return OGRE_NEW Entity(name, pMesh);
 
+    }
+   //-----------------------------------------------------------------------------
+    //-----------------------------------------------------------------------------
+    //-----------------------------------------------------------------------------
+    Entity::TempBlendedBufferInfo::~TempBlendedBufferInfo(void)
+    {
+        // check that temp buffers have been released
+        if (destPositionBuffer)
+            destPositionBuffer->getManager()->releaseVertexBufferCopy(destPositionBuffer);
+        if (destNormalBuffer)
+            destNormalBuffer->getManager()->releaseVertexBufferCopy(destNormalBuffer);
+
+    }
+    //-----------------------------------------------------------------------------
+    void Entity::TempBlendedBufferInfo::extractFrom(const VertexData* sourceData)
+    {
+        // Release old buffer copies first
+        if (destPositionBuffer)
+        {
+            destPositionBuffer->getManager()->releaseVertexBufferCopy(destPositionBuffer);
+            assert(!destPositionBuffer);
+        }
+        if (destNormalBuffer)
+        {
+            destNormalBuffer->getManager()->releaseVertexBufferCopy(destNormalBuffer);
+            assert(!destNormalBuffer);
+        }
+
+        VertexDeclaration* decl = sourceData->vertexDeclaration;
+        VertexBufferBinding* bind = sourceData->vertexBufferBinding;
+        const VertexElement *posElem = decl->findElementBySemantic(VES_POSITION);
+        const VertexElement *normElem = decl->findElementBySemantic(VES_NORMAL);
+
+        assert(posElem && "Positions are required");
+
+        posBindIndex = posElem->getSource();
+        srcPositionBuffer = bind->getBuffer(posBindIndex);
+        srcNormalBuffer.reset();
+
+        if (!normElem)
+        {
+            posNormalShareBuffer = false;
+            posNormalExtraData = posElem->getSize() != srcPositionBuffer->getVertexSize();
+        }
+        else
+        {
+            normBindIndex = normElem->getSource();
+            if (normBindIndex == posBindIndex)
+            {
+                posNormalShareBuffer = true;
+                posNormalExtraData = (posElem->getSize() + normElem->getSize()) != srcPositionBuffer->getVertexSize();
+            }
+            else
+            {
+                posNormalExtraData = false;
+                posNormalShareBuffer = false;
+                srcNormalBuffer = bind->getBuffer(normBindIndex);
+            }
+        }
+    }
+    //-----------------------------------------------------------------------------
+    void Entity::TempBlendedBufferInfo::checkoutTempCopies(bool positions, bool normals)
+    {
+        bindPositions = positions;
+        bindNormals = normals;
+
+        if (positions && !destPositionBuffer)
+        {
+            destPositionBuffer =
+                srcPositionBuffer->getManager()->allocateVertexBufferCopy(srcPositionBuffer, this, posNormalExtraData);
+        }
+        if (normals && !posNormalShareBuffer && srcNormalBuffer && !destNormalBuffer)
+        {
+            destNormalBuffer = srcNormalBuffer->getManager()->allocateVertexBufferCopy(srcNormalBuffer, this);
+        }
+    }
+    //-----------------------------------------------------------------------------
+    bool Entity::TempBlendedBufferInfo::buffersCheckedOut(bool positions, bool normals) const
+    {
+        if (positions || (normals && posNormalShareBuffer))
+        {
+            if (!destPositionBuffer)
+                return false;
+
+            destPositionBuffer->getManager()->touchVertexBufferCopy(destPositionBuffer);
+        }
+
+        if (normals && !posNormalShareBuffer)
+        {
+            if (!destNormalBuffer)
+                return false;
+
+            destNormalBuffer->getManager()->touchVertexBufferCopy(destNormalBuffer);
+        }
+
+        return true;
+    }
+    //-----------------------------------------------------------------------------
+    void Entity::TempBlendedBufferInfo::bindTempCopies(VertexData* targetData, bool suppressHardwareUpload)
+    {
+        this->destPositionBuffer->suppressHardwareUpdate(suppressHardwareUpload);
+        targetData->vertexBufferBinding->setBinding(
+            this->posBindIndex, this->destPositionBuffer);
+        if (bindNormals && !posNormalShareBuffer && destNormalBuffer)
+        {
+            this->destNormalBuffer->suppressHardwareUpdate(suppressHardwareUpload);
+            targetData->vertexBufferBinding->setBinding(
+                this->normBindIndex, this->destNormalBuffer);
+        }
+    }
+    //-----------------------------------------------------------------------------
+    void Entity::TempBlendedBufferInfo::licenseExpired(HardwareBuffer* buffer)
+    {
+        assert(buffer == destPositionBuffer.get()
+            || buffer == destNormalBuffer.get());
+
+        if (buffer == destPositionBuffer.get())
+            destPositionBuffer.reset();
+        if (buffer == destNormalBuffer.get())
+            destNormalBuffer.reset();
     }
 }
