@@ -55,8 +55,7 @@ namespace Ogre {
         mLinkedReflectPlane(0),
         mLinkedObliqueProjPlane(0),
         mReflect(false),
-        mObliqueDepthProjection(false),
-        mOrientationMode(OR_DEGREE_0)
+        mObliqueDepthProjection(false)
     {
         // Alter superclass members
         mVisible = false;
@@ -352,13 +351,6 @@ namespace Ogre {
     {
         // Common calcs
         RealRect rect = calcProjectionParameters();
-
-        if (!OGRE_NO_VIEWPORT_ORIENTATIONMODE && mOrientationMode != OR_PORTRAIT)
-        {
-            std::swap(rect.left, rect.bottom);
-            std::swap(rect.right, rect.top);
-        }
-
         Real left = rect.left, right = rect.right, top = rect.top, bottom = rect.bottom;
 
         if (!mCustomProjMatrix)
@@ -462,11 +454,6 @@ namespace Ogre {
             } // ortho            
         } // !mCustomProjMatrix
 
-#if OGRE_NO_VIEWPORT_ORIENTATIONMODE == 0
-        // Deal with orientation mode
-        mProjMatrix = mProjMatrix *
-            Matrix4(Quaternion(Degree(mOrientationMode * 90.f), Vector3::UNIT_Z));
-#endif
         RenderSystem* renderSystem = Root::getSingleton().getRenderSystem();
 
         if(renderSystem)
@@ -1128,25 +1115,6 @@ namespace Ogre {
         volume.planes.push_back(mFrustumPlanes[FRUSTUM_PLANE_LEFT]);
         volume.planes.push_back(mFrustumPlanes[FRUSTUM_PLANE_RIGHT]);
         return volume;
-    }
-    //---------------------------------------------------------------------
-    void Frustum::setOrientationMode(OrientationMode orientationMode)
-    {
-#if OGRE_NO_VIEWPORT_ORIENTATIONMODE != 0
-        OGRE_EXCEPT(Exception::ERR_NOT_IMPLEMENTED,
-                    "Setting Frustrum orientation mode is not supported");
-#endif
-        mOrientationMode = orientationMode;
-        invalidateFrustum();
-    }
-    //---------------------------------------------------------------------
-    OrientationMode Frustum::getOrientationMode() const
-    {
-#if OGRE_NO_VIEWPORT_ORIENTATIONMODE != 0
-        OGRE_EXCEPT(Exception::ERR_NOT_IMPLEMENTED,
-                    "Getting Frustrum orientation mode is not supported");
-#endif
-        return mOrientationMode;
     }
 
 
