@@ -142,19 +142,10 @@ namespace Ogre {
         mVertexBufferBindings.clear();
     }
     //-----------------------------------------------------------------------
-    void HardwareBufferManagerBase::registerVertexBufferSourceAndCopy(
-            const HardwareVertexBufferSharedPtr& sourceBuffer,
-            const HardwareVertexBufferSharedPtr& copy)
-    {
-            OGRE_LOCK_MUTEX(mTempBuffersMutex);
-        // Add copy to free temporary vertex buffers
-        mFreeTempVertexBufferMap.emplace(sourceBuffer.get(), copy);
-    }
-    //-----------------------------------------------------------------------
     HardwareVertexBufferSharedPtr 
     HardwareBufferManagerBase::allocateVertexBufferCopy(
         const HardwareVertexBufferSharedPtr& sourceBuffer, 
-        BufferLicenseType licenseType, HardwareBufferLicensee* licensee,
+        HardwareBufferLicensee* licensee,
         bool copyData)
     {
         // pre-lock the mVertexBuffers mutex, which would usually get locked in
@@ -171,12 +162,14 @@ namespace Ogre {
                 mFreeTempVertexBufferMap.find(sourceBuffer.get());
             if (i == mFreeTempVertexBufferMap.end())
             {
+                printf("creating temp vertex buffer\n");
                 // create new copy buffer, use shadow buffer and make dynamic
                 vbuf = createVertexBuffer(sourceBuffer->getVertexSize(), sourceBuffer->getNumVertices(), HBU_CPU_TO_GPU,
                                           true);
             }
             else
             {
+                printf("Reusing temp vertex buffer\n");
                 // Allocate existing copy
                 vbuf = i->second;
                 mFreeTempVertexBufferMap.erase(i);
@@ -191,7 +184,7 @@ namespace Ogre {
             // Insert copy into licensee list
             mTempVertexBufferLicenses.emplace(
                     vbuf.get(),
-                    VertexBufferLicense(sourceBuffer.get(), licenseType, EXPIRED_DELAY_FRAME_THRESHOLD, vbuf, licensee));
+                    VertexBufferLicense(sourceBuffer.get(), EXPIRED_DELAY_FRAME_THRESHOLD, vbuf, licensee));
             return vbuf;
         }
 
@@ -205,6 +198,7 @@ namespace Ogre {
             mTempVertexBufferLicenses.find(bufferCopy.get());
         if (i != mTempVertexBufferLicenses.end())
         {
+            printf("Releasing temp vertex buffer\n");
             const VertexBufferLicense& vbl = i->second;
             vbl.licensee->licenseExpired(vbl.buffer.get());
             mFreeTempVertexBufferMap.emplace(vbl.originalBufferPtr, vbl.buffer);
@@ -221,7 +215,6 @@ namespace Ogre {
         if (i != mTempVertexBufferLicenses.end())
         {
             VertexBufferLicense& vbl = i->second;
-            assert(vbl.licenseType == BLT_AUTOMATIC_RELEASE);
             vbl.expiredDelay = EXPIRED_DELAY_FRAME_THRESHOLD;
         }
     }
@@ -272,8 +265,7 @@ namespace Ogre {
         {
             TemporaryVertexBufferLicenseMap::iterator icur = i++;
             VertexBufferLicense& vbl = icur->second;
-            if (vbl.licenseType == BLT_AUTOMATIC_RELEASE &&
-                (forceFreeUnused || --vbl.expiredDelay <= 0))
+            if (forceFreeUnused || --vbl.expiredDelay <= 0)
             {
                 vbl.licensee->licenseExpired(vbl.buffer.get());
 
@@ -388,126 +380,5 @@ namespace Ogre {
                                                                      bool useShadowBuffer)
     {
         OGRE_EXCEPT(Exception::ERR_RENDERINGAPI_ERROR, "not supported by RenderSystem");
-    }
-    //-----------------------------------------------------------------------------
-    //-----------------------------------------------------------------------------
-    //-----------------------------------------------------------------------------
-    TempBlendedBufferInfo::~TempBlendedBufferInfo(void)
-    {
-        // check that temp buffers have been released
-        if (destPositionBuffer)
-            destPositionBuffer->getManager()->releaseVertexBufferCopy(destPositionBuffer);
-        if (destNormalBuffer)
-            destNormalBuffer->getManager()->releaseVertexBufferCopy(destNormalBuffer);
-
-    }
-    //-----------------------------------------------------------------------------
-    void TempBlendedBufferInfo::extractFrom(const VertexData* sourceData)
-    {
-        // Release old buffer copies first
-        if (destPositionBuffer)
-        {
-            destPositionBuffer->getManager()->releaseVertexBufferCopy(destPositionBuffer);
-            assert(!destPositionBuffer);
-        }
-        if (destNormalBuffer)
-        {
-            destNormalBuffer->getManager()->releaseVertexBufferCopy(destNormalBuffer);
-            assert(!destNormalBuffer);
-        }
-
-        VertexDeclaration* decl = sourceData->vertexDeclaration;
-        VertexBufferBinding* bind = sourceData->vertexBufferBinding;
-        const VertexElement *posElem = decl->findElementBySemantic(VES_POSITION);
-        const VertexElement *normElem = decl->findElementBySemantic(VES_NORMAL);
-
-        assert(posElem && "Positions are required");
-
-        posBindIndex = posElem->getSource();
-        srcPositionBuffer = bind->getBuffer(posBindIndex);
-        srcNormalBuffer.reset();
-
-        if (!normElem)
-        {
-            posNormalShareBuffer = false;
-            posNormalExtraData = posElem->getSize() != srcPositionBuffer->getVertexSize();
-        }
-        else
-        {
-            normBindIndex = normElem->getSource();
-            if (normBindIndex == posBindIndex)
-            {
-                posNormalShareBuffer = true;
-                posNormalExtraData = (posElem->getSize() + normElem->getSize()) != srcPositionBuffer->getVertexSize();
-            }
-            else
-            {
-                posNormalExtraData = false;
-                posNormalShareBuffer = false;
-                srcNormalBuffer = bind->getBuffer(normBindIndex);
-            }
-        }
-    }
-    //-----------------------------------------------------------------------------
-    void TempBlendedBufferInfo::checkoutTempCopies(bool positions, bool normals)
-    {
-        bindPositions = positions;
-        bindNormals = normals;
-
-        if (positions && !destPositionBuffer)
-        {
-            destPositionBuffer = srcPositionBuffer->getManager()->allocateVertexBufferCopy(srcPositionBuffer, 
-                HardwareBufferManagerBase::BLT_AUTOMATIC_RELEASE, this, posNormalExtraData);
-        }
-        if (normals && !posNormalShareBuffer && srcNormalBuffer && !destNormalBuffer)
-        {
-            destNormalBuffer = srcNormalBuffer->getManager()->allocateVertexBufferCopy(srcNormalBuffer, 
-                HardwareBufferManagerBase::BLT_AUTOMATIC_RELEASE, this);
-        }
-    }
-    //-----------------------------------------------------------------------------
-    bool TempBlendedBufferInfo::buffersCheckedOut(bool positions, bool normals) const
-    {
-        if (positions || (normals && posNormalShareBuffer))
-        {
-            if (!destPositionBuffer)
-                return false;
-
-            destPositionBuffer->getManager()->touchVertexBufferCopy(destPositionBuffer);
-        }
-
-        if (normals && !posNormalShareBuffer)
-        {
-            if (!destNormalBuffer)
-                return false;
-
-            destNormalBuffer->getManager()->touchVertexBufferCopy(destNormalBuffer);
-        }
-
-        return true;
-    }
-    //-----------------------------------------------------------------------------
-    void TempBlendedBufferInfo::bindTempCopies(VertexData* targetData, bool suppressHardwareUpload)
-    {
-        this->destPositionBuffer->suppressHardwareUpdate(suppressHardwareUpload);
-        targetData->vertexBufferBinding->setBinding(
-            this->posBindIndex, this->destPositionBuffer);
-        if (bindNormals && !posNormalShareBuffer && destNormalBuffer)
-        {
-            this->destNormalBuffer->suppressHardwareUpdate(suppressHardwareUpload);
-            targetData->vertexBufferBinding->setBinding(
-                this->normBindIndex, this->destNormalBuffer);
-        }
-    }
-    //-----------------------------------------------------------------------------
-    void TempBlendedBufferInfo::licenseExpired(HardwareBuffer* buffer)
-    {
-        assert(buffer == destPositionBuffer.get()
-            || buffer == destNormalBuffer.get());
-
-        if (buffer == destPositionBuffer.get())
-            destPositionBuffer.reset();
-        if (buffer == destNormalBuffer.get())
-            destNormalBuffer.reset();
     }
 }
