@@ -68,27 +68,14 @@ namespace Ogre {
         }
     }
     //-----------------------------------------------------------------------------
-    bool HardwareVertexBuffer::checkIfVertexInstanceDataIsSupported()
-    {
-        // Use the current render system
-        RenderSystem* rs = Root::getSingleton().getRenderSystem();
-
-        // Check if the supported  
-        return rs->getCapabilities()->hasCapability(RSC_VERTEX_BUFFER_INSTANCE_DATA);
-    }
-    //-----------------------------------------------------------------------------
     void HardwareVertexBuffer::setIsInstanceData( const bool val )
     {
-        if (val && !checkIfVertexInstanceDataIsSupported())
-        {
-            OGRE_EXCEPT(Exception::ERR_RENDERINGAPI_ERROR, 
-                "vertex instance data is not supported by the render system.", 
-                "HardwareVertexBuffer::checkIfInstanceDataSupported");
-        }
-        else
-        {
-            mIsInstanceData = val;  
-        }
+        RenderSystem* rs = Root::getSingleton().getRenderSystem();
+
+        OgreAssert(!val || rs->getCapabilities()->hasCapability(RSC_VERTEX_BUFFER_INSTANCE_DATA),
+                   "unsupported by rendersystem");
+
+        mIsInstanceData = val;
     }
     //-----------------------------------------------------------------------------
     size_t HardwareVertexBuffer::getInstanceDataStepRate() const
