@@ -1386,9 +1386,8 @@ namespace Ogre {
         }
     }
     //---------------------------------------------------------------------
-    void Mesh::buildTangentVectors(VertexElementSemantic targetSemantic, 
-        unsigned short sourceTexCoordSet, unsigned short index, 
-        bool splitMirrored, bool splitRotated, bool storeParityInW)
+    void Mesh::buildTangentVectors(unsigned short sourceTexCoordSet, bool splitMirrored, bool splitRotated,
+                                   bool storeParityInW)
     {
 
         TangentSpaceCalc tangentsCalc;
@@ -1411,8 +1410,7 @@ namespace Ogre {
             }
             if (found)
             {
-                TangentSpaceCalc::Result res = 
-                    tangentsCalc.build(targetSemantic, sourceTexCoordSet, index);
+                TangentSpaceCalc::Result res = tangentsCalc.build(sourceTexCoordSet);
 
                 // If any vertex splitting happened, we have to give them bone assignments
                 if (mSkeleton)
@@ -1461,8 +1459,7 @@ namespace Ogre {
                 tangentsCalc.clear();
                 tangentsCalc.setVertexData(sm->vertexData);
                 tangentsCalc.addIndexData(sm->indexData, sm->operationType);
-                TangentSpaceCalc::Result res = 
-                    tangentsCalc.build(targetSemantic, sourceTexCoordSet, index);
+                TangentSpaceCalc::Result res = tangentsCalc.build(sourceTexCoordSet);
 
                 // If any vertex splitting happened, we have to give them bone assignments
                 if (mSkeleton)
@@ -1490,8 +1487,7 @@ namespace Ogre {
 
     }
     //---------------------------------------------------------------------
-    bool Mesh::suggestTangentVectorBuildParams(VertexElementSemantic targetSemantic,
-        unsigned short& outSourceCoordSet, unsigned short& outIndex)
+    bool Mesh::suggestTangentVectorBuildParams(unsigned short& outSourceCoordSet)
     {
         // Go through all the vertex data and locate source and dest (must agree)
         bool sharedGeometryDone = false;
@@ -1516,58 +1512,31 @@ namespace Ogre {
             unsigned short targetIndex = 0;
             for (targetIndex = 0; targetIndex < OGRE_MAX_TEXTURE_COORD_SETS; ++targetIndex)
             {
-                const VertexElement* testElem =
-                    vertexData->vertexDeclaration->findElementBySemantic(
-                        VES_TEXTURE_COORDINATES, targetIndex);
+                auto testElem =
+                    vertexData->vertexDeclaration->findElementBySemantic(VES_TEXTURE_COORDINATES, targetIndex);
                 if (!testElem)
                     break; // finish if we've run out, t will be the target
 
-                if (!sourceElem)
+                // We're still looking for the source texture coords
+                if (testElem->getType() == VET_FLOAT2)
                 {
-                    // We're still looking for the source texture coords
-                    if (testElem->getType() == VET_FLOAT2)
-                    {
-                        // Ok, we found it
-                        sourceElem = testElem;
-                    }
+                    // Ok, we found it
+                    sourceElem = testElem;
+                    break;
                 }
-                
-                if(!foundExisting && targetSemantic == VES_TEXTURE_COORDINATES)
-                {
-                    // We're looking for the destination
-                    // Check to see if we've found a possible
-                    if (testElem->getType() == VET_FLOAT3)
-                    {
-                        // This is a 3D set, might be tangents
-                        foundExisting = true;
-                    }
-
-                }
-
-            }
-
-            if (!foundExisting && targetSemantic != VES_TEXTURE_COORDINATES)
-            {
-                targetIndex = 0;
-                // Look for existing semantic
-                const VertexElement* testElem =
-                    vertexData->vertexDeclaration->findElementBySemantic(
-                    targetSemantic, targetIndex);
-                if (testElem)
-                {
-                    foundExisting = true;
-                }
-
             }
 
             // After iterating, we should have a source and a possible destination (t)
             if (!sourceElem)
             {
                 OGRE_EXCEPT(Exception::ERR_ITEM_NOT_FOUND,
-                    "Cannot locate an appropriate 2D texture coordinate set for "
-                    "all the vertex data in this mesh to create tangents from. ",
-                    "Mesh::suggestTangentVectorBuildParams");
+                            "Cannot locate an appropriate 2D texture coordinate set for "
+                            "all the vertex data in this mesh to create tangents from. ");
             }
+
+            // Look for existing semantic
+            foundExisting = vertexData->vertexDeclaration->findElementBySemantic(VES_TANGENT);
+
             // Check that we agree with previous decisions, if this is not the
             // first one, and if we're not just using the existing one
             if (!firstOne && !foundExisting)
@@ -1575,24 +1544,14 @@ namespace Ogre {
                 if (sourceElem->getIndex() != outSourceCoordSet)
                 {
                     OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS,
-                        "Multiple sets of vertex data in this mesh disagree on "
-                        "the appropriate index to use for the source texture coordinates. "
-                        "This ambiguity must be rectified before tangents can be generated.",
-                        "Mesh::suggestTangentVectorBuildParams");
-                }
-                if (targetIndex != outIndex)
-                {
-                    OGRE_EXCEPT(Exception::ERR_INVALIDPARAMS,
-                        "Multiple sets of vertex data in this mesh disagree on "
-                        "the appropriate index to use for the target texture coordinates. "
-                        "This ambiguity must be rectified before tangents can be generated.",
-                        "Mesh::suggestTangentVectorBuildParams");
+                                "Multiple sets of vertex data in this mesh disagree on "
+                                "the appropriate index to use for the texture coordinates. "
+                                "This ambiguity must be rectified before tangents can be generated.");
                 }
             }
 
             // Otherwise, save this result
             outSourceCoordSet = sourceElem->getIndex();
-            outIndex = targetIndex;
 
             firstOne = false;
 
