@@ -43,445 +43,98 @@ namespace Ogre {
     {  
         assert( msSingleton );  return ( *msSingleton );  
     }
-    //-----------------------------------------------------------------------
-    /** Enumerates the type of requests */
-    enum RequestType
-    {
-        RT_INITIALISE_GROUP = 0,
-        RT_INITIALISE_ALL_GROUPS = 1,
-        RT_PREPARE_GROUP = 2,
-        RT_PREPARE_RESOURCE = 3,
-        RT_LOAD_GROUP = 4,
-        RT_LOAD_RESOURCE = 5,
-        RT_UNLOAD_GROUP = 6,
-        RT_UNLOAD_RESOURCE = 7
-    };
-    /** Encapsulates a queued request for the background queue */
-    struct ResourceRequest
-    {
-        RequestType type;
-        String resourceName;
-        ResourceHandle resourceHandle;
-        String resourceType;
-        String groupName;
-        bool isManual;
-        ManualResourceLoader* loader;
-        NameValuePairList* loadParams;
-        ResourceBackgroundQueue::Listener* listener;
-        BackgroundProcessResult result;
-    };
-    /// Struct that holds details of queued notifications
-    struct ResourceResponse
-    {
-        ResourceResponse(ResourcePtr r, const ResourceRequest& req)
-            : resource(r), request(req)
-        {}
-
-        ResourcePtr resource;
-        ResourceRequest request;
-    };
     //------------------------------------------------------------------------
-    ResourceBackgroundQueue::ResourceBackgroundQueue() : mWorkQueueChannel(0)
+    ResourceBackgroundQueue::ResourceBackgroundQueue() {}
+    ResourceBackgroundQueue::~ResourceBackgroundQueue() {}
+    //------------------------------------------------------------------------
+    std::future<void> ResourceBackgroundQueue::initialiseResourceGroup(const String& name)
     {
+        auto task = std::make_shared<std::packaged_task<void()>>(
+            [name]() { ResourceGroupManager::getSingleton().initialiseResourceGroup(name); });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
     }
     //------------------------------------------------------------------------
-    ResourceBackgroundQueue::~ResourceBackgroundQueue()
+    std::future<void> ResourceBackgroundQueue::initialiseAllResourceGroups()
     {
-        shutdown();
+        auto task = std::make_shared<std::packaged_task<void()>>(
+            []() { ResourceGroupManager::getSingleton().initialiseAllResourceGroups(); });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
+    }
+    //------------------------------------------------------------------------
+    std::future<void> ResourceBackgroundQueue::prepareResourceGroup(const String& name)
+    {
+        auto task = std::make_shared<std::packaged_task<void()>>(
+            [name]() { ResourceGroupManager::getSingleton().prepareResourceGroup(name); });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
+    }
+    //------------------------------------------------------------------------
+    std::future<void> ResourceBackgroundQueue::loadResourceGroup(const String& name)
+    {
+        auto task = std::make_shared<std::packaged_task<void()>>([name]()
+        {
+#   if OGRE_THREAD_SUPPORT == 1
+            ResourceGroupManager::getSingleton().loadResourceGroup(name);
+#   else
+            ResourceGroupManager::getSingleton().prepareResourceGroup(name);
+            Root::getSingleton().getWorkQueue()->addMainThreadTask(
+                [name]() { ResourceGroupManager::getSingleton().loadResourceGroup(name); });
+#   endif
+        });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
+    }
+    //------------------------------------------------------------------------
+    std::future<void> ResourceBackgroundQueue::prepare(const ResourcePtr& res)
+    {
+        auto task = std::make_shared<std::packaged_task<void()>>(
+            [res]()
+            {
+                res->prepare(true);
+                Root::getSingleton().getWorkQueue()->addMainThreadTask([res]() { res->_firePreparingComplete(); });
+            });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
+    }
+    //------------------------------------------------------------------------
+    std::future<void> ResourceBackgroundQueue::load(const ResourcePtr& res)
+    {
+        auto task = std::make_shared<std::packaged_task<void()>>(
+            [res]()
+            {
+#if OGRE_THREAD_SUPPORT == 1
+                res->load(true);
+                Root::getSingleton().getWorkQueue()->addMainThreadTask([res]() { res->_fireLoadingComplete(); });
+#else
+                res->prepare(true);
+                Root::getSingleton().getWorkQueue()->addMainThreadTask(
+                    [res]()
+                    {
+                        res->load(true);
+                        res->_fireLoadingComplete();
+                    });
+#endif
+            });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
     }
     //---------------------------------------------------------------------
-    void ResourceBackgroundQueue::initialise()
+    std::future<void> ResourceBackgroundQueue::unload(const ResourcePtr& res)
     {
-        WorkQueue* wq = Root::getSingleton().getWorkQueue();
-        mWorkQueueChannel = wq->getChannel("Ogre/ResourceBGQ");
-        wq->addResponseHandler(mWorkQueueChannel, this);
-        wq->addRequestHandler(mWorkQueueChannel, this);
+        auto task = std::make_shared<std::packaged_task<void()>>([res]() { res->unload(); });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
     }
     //---------------------------------------------------------------------
-    void ResourceBackgroundQueue::shutdown()
+    std::future<void> ResourceBackgroundQueue::unloadResourceGroup(const String& name)
     {
-        WorkQueue* wq = Root::getSingleton().getWorkQueue();
-        wq->abortRequestsByChannel(mWorkQueueChannel);
-        wq->removeRequestHandler(mWorkQueueChannel, this);
-        wq->removeResponseHandler(mWorkQueueChannel, this);
+        auto task = std::make_shared<std::packaged_task<void()>>(
+            [name]() { ResourceGroupManager::getSingleton().unloadResourceGroup(name); });
+        Root::getSingleton().getWorkQueue()->addTask([task]() { (*task)(); });
+        return task->get_future();
     }
-    //------------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::initialiseResourceGroup(
-        const String& name, ResourceBackgroundQueue::Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_INITIALISE_GROUP;
-        req.groupName = name;
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceGroupManager::getSingleton().initialiseResourceGroup(name);
-        return 0; 
-#endif
-    }
-    //------------------------------------------------------------------------
-    BackgroundProcessTicket 
-    ResourceBackgroundQueue::initialiseAllResourceGroups( 
-        ResourceBackgroundQueue::Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_INITIALISE_ALL_GROUPS;
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceGroupManager::getSingleton().initialiseAllResourceGroups();
-        return 0; 
-#endif
-    }
-    //------------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::prepareResourceGroup(
-        const String& name, ResourceBackgroundQueue::Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_PREPARE_GROUP;
-        req.groupName = name;
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceGroupManager::getSingleton().prepareResourceGroup(name);
-        return 0; 
-#endif
-    }
-    //------------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::loadResourceGroup(
-        const String& name, ResourceBackgroundQueue::Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_LOAD_GROUP;
-        req.groupName = name;
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceGroupManager::getSingleton().loadResourceGroup(name);
-        return 0; 
-#endif
-    }
-    //------------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::prepare(
-        const String& resType, const String& name, 
-        const String& group, bool isManual, 
-        ManualResourceLoader* loader, 
-        const NameValuePairList* loadParams, 
-        ResourceBackgroundQueue::Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_PREPARE_RESOURCE;
-        req.resourceType = resType;
-        req.resourceName = name;
-        req.groupName = group;
-        req.isManual = isManual;
-        req.loader = loader;
-        // Make instance copy of loadParams for thread independence
-        req.loadParams = ( loadParams ? OGRE_NEW_T(NameValuePairList, MEMCATEGORY_GENERAL)( *loadParams ) : 0 );
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceManager* rm = 
-            ResourceGroupManager::getSingleton()._getResourceManager(resType);
-        rm->prepare(name, group, isManual, loader, loadParams);
-        return 0; 
-#endif
-    }
-    //------------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::load(
-        const String& resType, const String& name, 
-        const String& group, bool isManual, 
-        ManualResourceLoader* loader, 
-        const NameValuePairList* loadParams, 
-        ResourceBackgroundQueue::Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_LOAD_RESOURCE;
-        req.resourceType = resType;
-        req.resourceName = name;
-        req.groupName = group;
-        req.isManual = isManual;
-        req.loader = loader;
-        // Make instance copy of loadParams for thread independence
-        req.loadParams = ( loadParams ? OGRE_NEW_T(NameValuePairList, MEMCATEGORY_GENERAL)( *loadParams ) : 0 );
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceManager* rm = 
-            ResourceGroupManager::getSingleton()._getResourceManager(resType);
-        rm->load(name, group, isManual, loader, loadParams);
-        return 0; 
-#endif
-    }
-    //---------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::unload(
-        const String& resType, const String& name, Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_UNLOAD_RESOURCE;
-        req.resourceType = resType;
-        req.resourceName = name;
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceManager* rm = 
-            ResourceGroupManager::getSingleton()._getResourceManager(resType);
-        rm->unload(name, ResourceGroupManager::AUTODETECT_RESOURCE_GROUP_NAME);
-        return 0; 
-#endif
-
-    }
-    //---------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::unload(
-        const String& resType, ResourceHandle handle, Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_UNLOAD_RESOURCE;
-        req.resourceType = resType;
-        req.resourceHandle = handle;
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceManager* rm = 
-            ResourceGroupManager::getSingleton()._getResourceManager(resType);
-        rm->unload(handle);
-        return 0; 
-#endif
-
-    }
-    //---------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::unloadResourceGroup(
-        const String& name, Listener* listener)
-    {
-#if OGRE_THREAD_SUPPORT
-        // queue a request
-        ResourceRequest req;
-        req.type = RT_UNLOAD_GROUP;
-        req.groupName = name;
-        req.listener = listener;
-        return addRequest(req);
-#else
-        // synchronous
-        ResourceGroupManager::getSingleton().unloadResourceGroup(name);
-        return 0; 
-#endif
-
-    }
-    //------------------------------------------------------------------------
-    bool ResourceBackgroundQueue::isProcessComplete(
-            BackgroundProcessTicket ticket)
-    {
-        return mOutstandingRequestSet.find(ticket) == mOutstandingRequestSet.end();
-    }
-    //------------------------------------------------------------------------
-    void ResourceBackgroundQueue::abortRequest( BackgroundProcessTicket ticket )
-    {
-        WorkQueue* queue = Root::getSingleton().getWorkQueue();
-
-        queue->abortRequest( ticket );
-    }
-    //------------------------------------------------------------------------
-    BackgroundProcessTicket ResourceBackgroundQueue::addRequest(ResourceRequest& req)
-    {
-        WorkQueue* queue = Root::getSingleton().getWorkQueue();
-
-        Any data(req);
-
-        WorkQueue::RequestID requestID = 
-            queue->addRequest(mWorkQueueChannel, (uint16)req.type, data);
-
-
-        mOutstandingRequestSet.insert(requestID);
-
-        return requestID;
-    }
-    //-----------------------------------------------------------------------
-    bool ResourceBackgroundQueue::canHandleRequest(const WorkQueue::Request* req, const WorkQueue* srcQ)
-    {
-        return true;
-    }
-    //-----------------------------------------------------------------------
-    WorkQueue::Response* ResourceBackgroundQueue::handleRequest(const WorkQueue::Request* req, const WorkQueue* srcQ)
-    {
-
-        ResourceRequest resreq = any_cast<ResourceRequest>(req->getData());
-
-        if( req->getAborted() )
-        {
-            if( resreq.type == RT_PREPARE_RESOURCE || resreq.type == RT_LOAD_RESOURCE )
-            {
-                OGRE_DELETE_T(resreq.loadParams, NameValuePairList, MEMCATEGORY_GENERAL);
-                resreq.loadParams = 0;
-            }
-            resreq.result.error = false;
-            ResourceResponse resresp(ResourcePtr(), resreq);
-            return OGRE_NEW WorkQueue::Response(req, true, resresp);
-        }
-
-        ResourceManager* rm = 0;
-        ResourcePtr resource;
-        try
-        {
-
-            switch (resreq.type)
-            {
-            case RT_INITIALISE_GROUP:
-                ResourceGroupManager::getSingleton().initialiseResourceGroup(
-                    resreq.groupName);
-                break;
-            case RT_INITIALISE_ALL_GROUPS:
-                ResourceGroupManager::getSingleton().initialiseAllResourceGroups();
-                break;
-            case RT_PREPARE_GROUP:
-                ResourceGroupManager::getSingleton().prepareResourceGroup(
-                    resreq.groupName);
-                break;
-            case RT_LOAD_GROUP:
-    #if OGRE_THREAD_SUPPORT == 2
-                ResourceGroupManager::getSingleton().prepareResourceGroup(
-                    resreq.groupName);
-    #else
-                ResourceGroupManager::getSingleton().loadResourceGroup(
-                    resreq.groupName);
-    #endif
-                break;
-            case RT_UNLOAD_GROUP:
-                ResourceGroupManager::getSingleton().unloadResourceGroup(
-                    resreq.groupName);
-                break;
-            case RT_PREPARE_RESOURCE:
-                rm = ResourceGroupManager::getSingleton()._getResourceManager(
-                    resreq.resourceType);
-                resource = rm->prepare(resreq.resourceName, resreq.groupName, resreq.isManual, 
-                    resreq.loader, resreq.loadParams, true);
-                break;
-            case RT_LOAD_RESOURCE:
-                rm = ResourceGroupManager::getSingleton()._getResourceManager(
-                    resreq.resourceType);
-    #if OGRE_THREAD_SUPPORT == 2
-                resource = rm->prepare(resreq.resourceName, resreq.groupName, resreq.isManual, 
-                    resreq.loader, resreq.loadParams, true);
-    #else
-                resource = rm->load(resreq.resourceName, resreq.groupName, resreq.isManual, 
-                    resreq.loader, resreq.loadParams, true);
-    #endif
-                break;
-            case RT_UNLOAD_RESOURCE:
-                rm = ResourceGroupManager::getSingleton()._getResourceManager(
-                    resreq.resourceType);
-                if (resreq.resourceName.empty())
-                    rm->unload(resreq.resourceHandle);
-                else
-                    rm->unload(resreq.resourceName, ResourceGroupManager::AUTODETECT_RESOURCE_GROUP_NAME);
-                break;
-            };
-        }
-        catch (Exception& e)
-        {
-            if( resreq.type == RT_PREPARE_RESOURCE || resreq.type == RT_LOAD_RESOURCE )
-            {
-                OGRE_DELETE_T(resreq.loadParams, NameValuePairList, MEMCATEGORY_GENERAL);
-                resreq.loadParams = 0;
-            }
-            resreq.result.error = true;
-            resreq.result.message = e.getFullDescription();
-
-            // return error response
-            ResourceResponse resresp(resource, resreq);
-            return OGRE_NEW WorkQueue::Response(req, false, resresp, e.getFullDescription());
-        }
-
-
-        // success
-        if( resreq.type == RT_PREPARE_RESOURCE || resreq.type == RT_LOAD_RESOURCE )
-        {
-            OGRE_DELETE_T(resreq.loadParams, NameValuePairList, MEMCATEGORY_GENERAL);
-            resreq.loadParams = 0;
-        }
-        resreq.result.error = false;
-        ResourceResponse resresp(resource, resreq);
-        return OGRE_NEW WorkQueue::Response(req, true, resresp);
-
-    }
-    //------------------------------------------------------------------------
-    bool ResourceBackgroundQueue::canHandleResponse(const WorkQueue::Response* res, const WorkQueue* srcQ)
-    {
-        return true;
-    }
-    //------------------------------------------------------------------------
-    void ResourceBackgroundQueue::handleResponse(const WorkQueue::Response* res, const WorkQueue* srcQ)
-    {
-        if( res->getRequest()->getAborted() )
-        {
-            mOutstandingRequestSet.erase(res->getRequest()->getID());
-            return ;
-        }
-
-        ResourceResponse resresp = any_cast<ResourceResponse>(res->getData());
-
-        // Complete full loading in main thread if semithreading
-        const ResourceRequest& req = resresp.request;
-
-        if (res->succeeded())
-        {
-#if OGRE_THREAD_SUPPORT == 2
-            // These load commands would have been downgraded to prepare() for the background
-            if (req.type == RT_LOAD_RESOURCE)
-            {
-                ResourceManager *rm = ResourceGroupManager::getSingleton()
-                    ._getResourceManager(req.resourceType);
-                rm->load(req.resourceName, req.groupName, req.isManual, req.loader, req.loadParams, true);
-            } 
-            else if (req.type == RT_LOAD_GROUP) 
-            {
-                ResourceGroupManager::getSingleton().loadResourceGroup(req.groupName);
-            }
-#endif
-            mOutstandingRequestSet.erase(res->getRequest()->getID());
-
-            // Call resource listener
-            if (resresp.resource) 
-            {
-
-                if (req.type == RT_LOAD_RESOURCE) 
-                {
-                    resresp.resource->_fireLoadingComplete();
-                } 
-                else 
-                {
-                    resresp.resource->_firePreparingComplete();
-                }
-            }
-        }
-        // Call queue listener
-        if (req.listener)
-            req.listener->operationCompleted(res->getRequest()->getID(), req.result);
-    }
-    //------------------------------------------------------------------------
 
 }
 
