@@ -748,7 +748,7 @@ namespace Ogre {
             HardwareBufferManager::getSingleton()._releaseBufferCopies();
 
         // Tell the queue to process responses
-        mWorkQueue->processResponses();
+        mWorkQueue->processMainThreadTasks();
 
         OgreProfileEndGroup("Frame", OGREPROF_GENERAL);
 
@@ -885,7 +885,6 @@ namespace Ogre {
 
         // Since background thread might be access resources,
         // ensure shutdown before destroying resource manager.
-        mResourceBackgroundQueue->shutdown();
         mWorkQueue->shutdown();
 
         if(mSceneManagerEnum)
@@ -1203,7 +1202,6 @@ namespace Ogre {
         mActiveRenderer->getCapabilities()->log(LogManager::getSingleton().getDefaultLog());
 
         // Background loader
-        mResourceBackgroundQueue->initialise();
         mWorkQueue->startup();
         // Initialise material manager
         mMaterialManager->initialise();
