@@ -30,17 +30,11 @@ THE SOFTWARE.
 #include "OgreTimer.h"
 
 namespace Ogre {
-    //---------------------------------------------------------------------
-    uint16 WorkQueue::getChannel(const String& channelName)
+    void WorkQueue::processMainThreadTasks()
     {
-            OGRE_WQ_LOCK_MUTEX(mChannelMapMutex);
-
-        ChannelMap::iterator i = mChannelMap.find(channelName);
-        if (i == mChannelMap.end())
-        {
-            i = mChannelMap.insert(ChannelMap::value_type(channelName, mNextChannel++)).first;
-        }
-        return i->second;
+        OGRE_IGNORE_DEPRECATED_BEGIN
+        processResponses();
+        OGRE_IGNORE_DEPRECATED_END
     }
     //---------------------------------------------------------------------
     WorkQueue::Request::Request(uint16 channel, uint16 rtype, const Any& rData, uint8 retry, RequestID rid)
@@ -73,13 +67,9 @@ namespace Ogre {
         , mWorkerRenderSystemAccess(false)
         , mIsRunning(false)
         , mResposeTimeLimitMS(10)
-        , mWorkerFunc(0)
-        , mRequestCount(0)
         , mPaused(false)
         , mAcceptRequests(true)
         , mShuttingDown(false)
-        , mIdleThreadRunning(false)
-        , mIdleProcessed(0)
     {
     }
     //---------------------------------------------------------------------
@@ -111,363 +101,29 @@ namespace Ogre {
     DefaultWorkQueueBase::~DefaultWorkQueueBase()
     {
         //shutdown(); // can't call here; abstract function
-
-        for (RequestQueue::iterator i = mRequestQueue.begin(); i != mRequestQueue.end(); ++i)
-        {
-            OGRE_DELETE (*i);
-        }
-        mRequestQueue.clear();
-
-        for (ResponseQueue::iterator i = mResponseQueue.begin(); i != mResponseQueue.end(); ++i)
-        {
-            OGRE_DELETE (*i);
-        }
-        mResponseQueue.clear();
+    }
+    void DefaultWorkQueueBase::addMainThreadTask(std::function<void()> task)
+    {
+        if (!mAcceptRequests || mShuttingDown)
+            return;
+        OGRE_WQ_LOCK_MUTEX(mResponseMutex);
+        mMainThreadTasks.push_back(task);
     }
     //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::addRequestHandler(uint16 channel, RequestHandler* rh)
+    void DefaultWorkQueueBase::addTask(std::function<void()> task)
     {
-            OGRE_WQ_LOCK_RW_MUTEX_WRITE(mRequestHandlerMutex);
-
-        RequestHandlerListByChannel::iterator i = mRequestHandlers.emplace(channel, RequestHandlerList()).first;
-
-        RequestHandlerList& handlers = i->second;
-        bool duplicate = false;
-        for (RequestHandlerList::iterator j = handlers.begin(); j != handlers.end(); ++j)
-        {
-            if ((*j)->getHandler() == rh)
-            {
-                duplicate = true;
-                break;
-            }
-        }
-        if (!duplicate)
-            handlers.push_back(RequestHandlerHolderPtr(OGRE_NEW RequestHandlerHolder(rh)));
-
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::removeRequestHandler(uint16 channel, RequestHandler* rh)
-    {
-        OGRE_WQ_LOCK_RW_MUTEX_WRITE(mRequestHandlerMutex);
-
-        RequestHandlerListByChannel::iterator i = mRequestHandlers.find(channel);
-        if (i != mRequestHandlers.end())
-        {
-            RequestHandlerList& handlers = i->second;
-            for (RequestHandlerList::iterator j = handlers.begin(); j != handlers.end(); ++j)
-            {
-                if ((*j)->getHandler() == rh)
-                {
-                    // Disconnect - this will make it safe across copies of the list
-                    // this is threadsafe and will wait for existing processes to finish
-                    (*j)->disconnectHandler();
-                    handlers.erase(j);  
-                    break;
-                }
-            }
-
-        }
-
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::addResponseHandler(uint16 channel, ResponseHandler* rh)
-    {
-        ResponseHandlerListByChannel::iterator i = mResponseHandlers.emplace(channel, ResponseHandlerList()).first;
-
-        ResponseHandlerList& handlers = i->second;
-        if (std::find(handlers.begin(), handlers.end(), rh) == handlers.end())
-            handlers.push_back(rh);
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::removeResponseHandler(uint16 channel, ResponseHandler* rh)
-    {
-        ResponseHandlerListByChannel::iterator i = mResponseHandlers.find(channel);
-        if (i != mResponseHandlers.end())
-        {
-            ResponseHandlerList& handlers = i->second;
-            ResponseHandlerList::iterator j = std::find(
-                handlers.begin(), handlers.end(), rh);
-            if (j != handlers.end())
-                handlers.erase(j);
-
-        }
-    }
-    //---------------------------------------------------------------------
-    WorkQueue::RequestID DefaultWorkQueueBase::addRequest(uint16 channel, uint16 requestType, 
-        const Any& rData, uint8 retryCount, bool forceSynchronous, bool idleThread)
-    {
-        Request* req = 0;
-        RequestID rid = 0;
-
-        {
-            // lock to acquire rid and push request to the queue
-                    OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-
-            if (!mAcceptRequests || mShuttingDown)
-                return 0;
-
-            rid = ++mRequestCount;
-            req = OGRE_NEW Request(channel, requestType, rData, retryCount, rid);
-
-            LogManager::getSingleton().stream(LML_TRIVIAL) << 
-                "DefaultWorkQueueBase('" << mName << "') - QUEUED(thread:" <<
-                OGRE_THREAD_CURRENT_ID
-                << "): ID=" << rid
-                << " channel=" << channel << " requestType=" << requestType;
-#if OGRE_THREAD_SUPPORT
-            if (!forceSynchronous&& !idleThread)
-            {
-                mRequestQueue.push_back(req);
-                notifyWorkers();
-                return rid;
-            }
-#endif
-        }
-        if(OGRE_THREAD_SUPPORT && idleThread){
-            OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-            mIdleRequestQueue.push_back(req);
-            if(!mIdleThreadRunning)
-            {
-                notifyWorkers();
-            }
-        } else { //forceSynchronous
-            processRequestResponse(req, true);
-        }
-        return rid;
-
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::addRequestWithRID(WorkQueue::RequestID rid, uint16 channel, 
-        uint16 requestType, const Any& rData, uint8 retryCount)
-    {
-        // lock to push request to the queue
-            OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-
-        if (mShuttingDown)
+        OGRE_WQ_LOCK_MUTEX(mRequestMutex);
+        if (!mAcceptRequests || mShuttingDown)
             return;
 
-        Request* req = OGRE_NEW Request(channel, requestType, rData, retryCount, rid);
-
-        LogManager::getSingleton().stream(LML_TRIVIAL) << 
-            "DefaultWorkQueueBase('" << mName << "') - REQUEUED(thread:" <<
-            OGRE_THREAD_CURRENT_ID
-            << "): ID=" << rid
-                   << " channel=" << channel << " requestType=" << requestType;
 #if OGRE_THREAD_SUPPORT
-        mRequestQueue.push_back(req);
+        mTasks.push_back(task);
         notifyWorkers();
 #else
-        processRequestResponse(req, true);
+        task(); // no threading, just run it
 #endif
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::abortRequest(RequestID id)
-    {
-            OGRE_WQ_LOCK_MUTEX(mProcessMutex);
-
-        // NOTE: Pending requests are exist any of RequestQueue, ProcessQueue and
-        // ResponseQueue when keeping ProcessMutex, so we check all of these queues.
-
-        for (RequestQueue::iterator i = mProcessQueue.begin(); i != mProcessQueue.end(); ++i)
-        {
-            if ((*i)->getID() == id)
-            {
-                (*i)->abortRequest();
-                break;
-            }
-        }
-
-        {
-                    OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-
-            for (RequestQueue::iterator i = mRequestQueue.begin(); i != mRequestQueue.end(); ++i)
-            {
-                if ((*i)->getID() == id)
-                {
-                    (*i)->abortRequest();
-                    break;
-                }
-            }
-        }
-
-        {
-            if(mIdleProcessed)
-            {
-                mIdleProcessed->abortRequest();
-            }
-
-            OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-            for (RequestQueue::iterator i = mIdleRequestQueue.begin(); i != mIdleRequestQueue.end(); ++i)
-            {
-                (*i)->abortRequest();
-            }
-        }
-
-        {
-                    OGRE_WQ_LOCK_MUTEX(mResponseMutex);
-
-            for (ResponseQueue::iterator i = mResponseQueue.begin(); i != mResponseQueue.end(); ++i)
-            {
-                if( (*i)->getRequest()->getID() == id )
-                {
-                    (*i)->abortRequest();
-                    break;
-                }
-            }
-        }
-    }
-    //---------------------------------------------------------------------
-    bool DefaultWorkQueueBase::abortPendingRequest(RequestID id)
-    {
-        // Request should not exist in idle queue and request queue simultaneously.
-        {
-            OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-            for (RequestQueue::iterator i = mRequestQueue.begin(); i != mRequestQueue.end(); ++i)
-            {
-                if ((*i)->getID() == id)
-                {
-                    (*i)->abortRequest();
-                    return true;
-                }
-            }
-        }
-        {
-            OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-            for (RequestQueue::iterator i = mIdleRequestQueue.begin(); i != mIdleRequestQueue.end(); ++i)
-            {
-                if ((*i)->getID() == id)
-                {
-                    (*i)->abortRequest();
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::abortRequestsByChannel(uint16 channel)
-    {
-            OGRE_WQ_LOCK_MUTEX(mProcessMutex);
-
-        for (RequestQueue::iterator i = mProcessQueue.begin(); i != mProcessQueue.end(); ++i)
-        {
-            if ((*i)->getChannel() == channel)
-            {
-                (*i)->abortRequest();
-            }
-        }
-
-        {
-                    OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-
-            for (RequestQueue::iterator i = mRequestQueue.begin(); i != mRequestQueue.end(); ++i)
-            {
-                if ((*i)->getChannel() == channel)
-                {
-                    (*i)->abortRequest();
-                }
-            }
-        }
-        {
-            if (mIdleProcessed && mIdleProcessed->getChannel() == channel)
-            {
-                mIdleProcessed->abortRequest();
-            }
-
-            OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-
-            for (RequestQueue::iterator i = mIdleRequestQueue.begin(); i != mIdleRequestQueue.end(); ++i)
-            {
-                if ((*i)->getChannel() == channel)
-                {
-                    (*i)->abortRequest();
-                }
-            }
-        }
-
-        {
-                    OGRE_WQ_LOCK_MUTEX(mResponseMutex);
-
-            for (ResponseQueue::iterator i = mResponseQueue.begin(); i != mResponseQueue.end(); ++i)
-            {
-                if( (*i)->getRequest()->getChannel() == channel )
-                {
-                    (*i)->abortRequest();
-                }
-            }
-        }
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::abortPendingRequestsByChannel(uint16 channel)
-    {
-        {
-                    OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-            for (RequestQueue::iterator i = mRequestQueue.begin(); i != mRequestQueue.end(); ++i)
-            {
-                if ((*i)->getChannel() == channel)
-                {
-                    (*i)->abortRequest();
-                }
-            }
-        }
-        {
-                    OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-
-                for (RequestQueue::iterator i = mIdleRequestQueue.begin(); i != mIdleRequestQueue.end(); ++i)
-                {
-                    if ((*i)->getChannel() == channel)
-                    {
-                        (*i)->abortRequest();
-                    }
-                }
-        }
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::abortAllRequests()
-    {
-            OGRE_WQ_LOCK_MUTEX(mProcessMutex);
-        {
-            for (RequestQueue::iterator i = mProcessQueue.begin(); i != mProcessQueue.end(); ++i)
-            {
-                (*i)->abortRequest();
-            }
-        }
-        
-
-        {
-                    OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-
-            for (RequestQueue::iterator i = mRequestQueue.begin(); i != mRequestQueue.end(); ++i)
-            {
-                (*i)->abortRequest();
-            }
-        }
-
-        {
-
-            if(mIdleProcessed)
-            {
-                mIdleProcessed->abortRequest();
-            }
-
-            OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-
-            for (RequestQueue::iterator i = mIdleRequestQueue.begin(); i != mIdleRequestQueue.end(); ++i)
-            {
-                (*i)->abortRequest();
-            }
-        }
-
-        {
-                    OGRE_WQ_LOCK_MUTEX(mResponseMutex);
-
-            for (ResponseQueue::iterator i = mResponseQueue.begin(); i != mResponseQueue.end(); ++i)
-            {
-                (*i)->abortRequest();
-            }
-        }
-
+        LogManager::getSingleton().stream(LML_TRIVIAL)
+            << "DefaultWorkQueueBase('" << mName << "') - QUEUED(thread:" << OGRE_THREAD_CURRENT_ID << ")";
     }
     //---------------------------------------------------------------------
     void DefaultWorkQueueBase::setPaused(bool pause)
@@ -496,103 +152,23 @@ namespace Ogre {
     //---------------------------------------------------------------------
     void DefaultWorkQueueBase::_processNextRequest()
     {
-        if(processIdleRequests()){
-            // Found idle requests.
+        if (mTasks.empty())
             return;
-        }
-        Request* request = 0;
+
+        std::function<void()> task;
         {
             // scoped to only lock while retrieving the next request
-                    OGRE_WQ_LOCK_MUTEX(mProcessMutex);
-            {
-                            OGRE_WQ_LOCK_MUTEX(mRequestMutex);
-
-                if (!mRequestQueue.empty())
-                {
-                    request = mRequestQueue.front();
-                    mRequestQueue.pop_front();
-                    mProcessQueue.push_back( request );
-                }
-            }
+            OGRE_WQ_LOCK_MUTEX(mRequestMutex);
+            LogManager::getSingleton().stream(LML_TRIVIAL)
+                << "DefaultWorkQueueBase('" << mName << "') - PROCESS_TASK(thread:" << OGRE_THREAD_CURRENT_ID
+                << ")";
+            task = std::move(mTasks.front());
+            mTasks.pop_front();
         }
-
-        if (request)
-        {
-            processRequestResponse(request, false);
-        }
-
-
+        task();
     }
     //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::processRequestResponse(Request* r, bool synchronous)
-    {
-        Response* response = processRequest(r);
-
-        OGRE_WQ_LOCK_MUTEX(mProcessMutex);
-
-        RequestQueue::iterator it;
-        for( it = mProcessQueue.begin(); it != mProcessQueue.end(); ++it )
-        {
-            if( (*it) == r )
-            {
-                mProcessQueue.erase( it );
-                break;
-            }
-        }
-        if( mIdleProcessed == r )
-        {
-            mIdleProcessed = 0;
-        }
-        if (response)
-        {
-            if (!response->succeeded())
-            {
-                // Failed, should we retry?
-                const Request* req = response->getRequest();
-                if (req->getRetryCount())
-                {
-                    addRequestWithRID(req->getID(), req->getChannel(), req->getType(), req->getData(), 
-                        req->getRetryCount() - 1);
-                    // discard response (this also deletes request)
-                    OGRE_DELETE response;
-                    return;
-                }
-            }
-            if (synchronous)
-            {
-                processResponse(response);
-                OGRE_DELETE response;
-            }
-            else
-            {
-                if( response->getRequest()->getAborted() )
-                {
-                    // destroy response user data
-                    response->abortRequest();
-                }
-                // Queue response
-                OGRE_WQ_LOCK_MUTEX(mResponseMutex);
-                mResponseQueue.push_back(response);
-                // no need to wake thread, this is processed by the main thread
-            }
-
-        }
-        else
-        {
-            if (!r->getAborted())
-            {
-            // no response, delete request
-            LogManager::getSingleton().stream(LML_WARNING) <<
-                "DefaultWorkQueueBase('" << mName << "') warning: no handler processed request "
-                << r->getID() << ", channel " << r->getChannel()
-                << ", type " << r->getType();
-            }
-            OGRE_DELETE r;
-        }
-
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::processResponses() 
+    void DefaultWorkQueueBase::processMainThreadTasks()
     {
         unsigned long msStart = Root::getSingleton().getTimer()->getMilliseconds();
         unsigned long msCurrent = 0;
@@ -600,25 +176,17 @@ namespace Ogre {
         // keep going until we run out of responses or out of time
         while(true)
         {
-            Response* response = 0;
+            if(!mMainThreadTasks.empty())
             {
-                            OGRE_WQ_LOCK_MUTEX(mResponseMutex);
-
-                if (mResponseQueue.empty())
-                    break; // exit loop
-                else
+                std::function<void()> task;
                 {
-                    response = mResponseQueue.front();
-                    mResponseQueue.pop_front();
+                    OGRE_WQ_LOCK_MUTEX(mResponseMutex);
+                    LogManager::getSingleton().stream(LML_TRIVIAL)
+                        << "DefaultWorkQueueBase('" << mName << "') - PROCESS_MAIN_TASK";
+                    task = std::move(mMainThreadTasks.front());
+                    mMainThreadTasks.pop_front();
                 }
-            }
-
-            if (response)
-            {
-                processResponse(response);
-
-                OGRE_DELETE response;
-
+                task();
             }
 
             // time limit
@@ -628,117 +196,6 @@ namespace Ogre {
                 if (msCurrent - msStart > mResposeTimeLimitMS)
                     break;
             }
-        }
-    }
-    //---------------------------------------------------------------------
-    WorkQueue::Response* DefaultWorkQueueBase::processRequest(Request* r)
-    {
-        RequestHandlerListByChannel handlerListCopy;
-        {
-            // lock the list only to make a copy of it, to maximise parallelism
-                    OGRE_WQ_LOCK_RW_MUTEX_READ(mRequestHandlerMutex);
-            
-            handlerListCopy = mRequestHandlers;
-            
-        }
-
-        Response* response = 0;
-
-        StringStream dbgMsg;
-        dbgMsg << "thread:" << OGRE_THREAD_CURRENT_ID
-            << "): ID=" << r->getID() << " channel=" << r->getChannel() 
-            << " requestType=" << r->getType();
-
-        LogManager::getSingleton().stream(LML_TRIVIAL) << 
-            "DefaultWorkQueueBase('" << mName << "') - PROCESS_REQUEST_START(" << dbgMsg.str();
-
-        RequestHandlerListByChannel::iterator i = handlerListCopy.find(r->getChannel());
-        if (i != handlerListCopy.end())
-        {
-            RequestHandlerList& handlers = i->second;
-            for (RequestHandlerList::reverse_iterator j = handlers.rbegin(); j != handlers.rend(); ++j)
-            {
-                // threadsafe call which tests canHandleRequest and calls it if so 
-                response = (*j)->handleRequest(r, this);
-
-                if (response)
-                    break;
-            }
-        }
-
-        LogManager::getSingleton().stream(LML_TRIVIAL) << 
-            "DefaultWorkQueueBase('" << mName << "') - PROCESS_REQUEST_END(" << dbgMsg.str()
-            << " processed=" << (response!=0);
-
-        return response;
-
-    }
-    //---------------------------------------------------------------------
-    void DefaultWorkQueueBase::processResponse(Response* r)
-    {
-        LogManager::getSingleton().stream(LML_TRIVIAL)
-            << "DefaultWorkQueueBase('" << mName << "') - PROCESS_RESPONSE: "
-            << "ID=" << r->getRequest()->getID() << " success=" << r->succeeded() << " messages=[" << r->getMessages()
-            << "] channel=" << r->getRequest()->getChannel() << " requestType=" << r->getRequest()->getType();
-
-        ResponseHandlerListByChannel::iterator i = mResponseHandlers.find(r->getRequest()->getChannel());
-        if (i != mResponseHandlers.end())
-        {
-            ResponseHandlerList& handlers = i->second;
-            for (ResponseHandlerList::reverse_iterator j = handlers.rbegin(); j != handlers.rend(); ++j)
-            {
-                if ((*j)->canHandleResponse(r, this))
-                {
-                    (*j)->handleResponse(r, this);
-                }
-            }
-        }
-    }
-
-    bool DefaultWorkQueueBase::processIdleRequests()
-    {
-        {
-                    OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-            if(mIdleRequestQueue.empty() || mIdleThreadRunning){
-                return false;
-            } else {
-                mIdleThreadRunning = true;
-            }
-        }
-        try {
-            while(1){
-                {
-                                    OGRE_WQ_LOCK_MUTEX(mProcessMutex); // mProcessMutex needs to be the top mutex to prevent livelocks
-                    {
-                                            OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-                        if(!mIdleRequestQueue.empty()){
-                            mIdleProcessed = mIdleRequestQueue.front();
-                            mIdleRequestQueue.pop_front();
-                        } else {
-                            mIdleProcessed = 0;
-                            mIdleThreadRunning = false;
-                            return true;
-                        }
-                    }
-                }
-                processRequestResponse(mIdleProcessed, false);
-            }
-        } catch (...) { // Normally this should not happen.
-            {
-                // It is very important to clean up or the idle thread will be locked forever!
-                            OGRE_WQ_LOCK_MUTEX(mProcessMutex);
-                {
-                                    OGRE_WQ_LOCK_MUTEX(mIdleMutex);
-                    if(mIdleProcessed){
-                        mIdleProcessed->abortRequest();
-                    }
-                    mIdleProcessed = 0;
-                    mIdleThreadRunning = false;
-                }
-            }
-            Ogre::LogManager::getSingleton().stream() << "Exception caught in top of worker thread!";
-
-            return true;
         }
     }
 }
