@@ -28,20 +28,11 @@ THE SOFTWARE.
 #ifndef __GLRenderToVertexBuffer_H__
 #define __GLRenderToVertexBuffer_H__
 
-#include "OgreRenderToVertexBuffer.h"
+#include "OgreGLRenderToVertexBufferCommon.h"
 #include "OgreGLPrerequisites.h"
 
 namespace Ogre {
-    /**
-        An object which renders geometry to a vertex.
-
-        This is especially useful together with geometry shaders, as you can
-        render procedural geometry which will get saved to a vertex buffer for
-        reuse later, without regenerating it again. You can also create shaders
-        that run on previous results of those shaders, creating stateful 
-        shaders.
-    */
-    class _OgreGLExport GLRenderToVertexBuffer : public RenderToVertexBuffer
+    class GLRenderToVertexBuffer : public GLRenderToVertexBufferCommon
     {    
     public:
         /** C'tor */
@@ -50,21 +41,12 @@ namespace Ogre {
         virtual ~GLRenderToVertexBuffer();
 
         /**
-            Get the render operation for this buffer 
-        */
-        void getRenderOperation(RenderOperation& op) override;
-
-        /**
             Update the contents of this vertex buffer by rendering
         */
         void update(SceneManager* sceneMgr) override;
     protected:
-        void reallocateBuffer(size_t index);
         void bindVerticesOutput(Pass* pass);
         GLint getGLSemanticType(VertexElementSemantic semantic);
-        String getSemanticVaryingName(VertexElementSemantic semantic, unsigned short index);
-        HardwareVertexBufferSharedPtr mVertexBuffers[2];
-        size_t mFrontBufferIndex;
         GLuint mPrimitivesDrawnQuery;
     };
 }
