@@ -29,13 +29,10 @@ Copyright (c) 2000-2014 Torus Knot Software Ltd
 #include "OgreGL3PlusRenderToVertexBuffer.h"
 #include "OgreHardwareBufferManager.h"
 #include "OgreRenderable.h"
-#include "OgreSceneManager.h"
 #include "OgreRoot.h"
 #include "OgreRenderSystem.h"
 #include "OgreGLSLProgramManager.h"
 #include "OgreGLSLShader.h"
-#include "OgreStringConverter.h"
-#include "OgreTechnique.h"
 
 namespace Ogre {
 
@@ -78,32 +75,6 @@ namespace Ogre {
         }
     }
 
-
-    static GLint getVertexCountPerPrimitive(RenderOperation::OperationType operationType)
-    {
-        // We can only get points, lines or triangles since they are the only
-        // legal R2VB output primitive types.
-        switch (operationType)
-        {
-        case RenderOperation::OT_POINT_LIST:
-            return 1;
-        case RenderOperation::OT_LINE_LIST:
-            return 2;
-        default:
-        case RenderOperation::OT_TRIANGLE_LIST:
-            return 3;
-        }
-    }
-
-
-    void GL3PlusRenderToVertexBuffer::getRenderOperation(RenderOperation& op)
-    {
-        op.operationType = mOperationType;
-        op.useIndexes = false;
-        op.vertexData = mVertexData.get();
-    }
-
-
     void GL3PlusRenderToVertexBuffer::bindVerticesOutput(Pass* pass)
     {
         VertexDeclaration* declaration = mVertexData->vertexDeclaration;
@@ -111,21 +82,6 @@ namespace Ogre {
 
         if (elemCount == 0)
             return;
-
-        // Store the output in a buffer.  The buffer has the same
-        // structure as the shader output vertex data.
-        // Note: 64 is the minimum number of interleaved
-        // attributes allowed by GL_EXT_transform_feedback so we
-        // are using it. Otherwise we could query during
-        // rendersystem initialisation and use a dynamic sized
-        // array.  But that would require C99.
-        size_t sourceBufferIndex = mTargetBufferIndex == 0 ? 1 : 0;
-
-        // Bind and fill vertex arrays + buffers.
-        reallocateBuffer(sourceBufferIndex);
-        reallocateBuffer(mTargetBufferIndex);
-        // GL3PlusHardwareVertexBuffer* sourceVertexBuffer = static_cast<GL3PlusHardwareVertexBuffer*>(mVertexBuffers[mSourceBufferIndex].get());
-        // GL3PlusHardwareVertexBuffer* targetVertexBuffer = static_cast<GL3PlusHardwareVertexBuffer*>(mVertexBuffers[mTargetBufferIndex].get());
 
         //TODO GL4+ glBindTransformFeedback
 
@@ -154,64 +110,26 @@ namespace Ogre {
         //         mResetRequested = true;
         //     }
 
-        //     if (mResetRequested || mResetsEveryUpdate)
-        //     {
-        //         // Use source data to render to first buffer.
-        //         mSourceRenderable->getRenderOperation(renderOp);
-        //         targetBufferIndex = 0;
-        //     }
-        //     else
-        //     {
-        //         // Use current front buffer to render to back buffer.
-        //         this->getRenderOperation(renderOp);
-        //         targetBufferIndex = 1 - mSourceBufferIndex;
-        //     }
+        Ogre::Pass* r2vbPass = derivePass(sceneMgr);
 
-        //     if (!mVertexBuffers[targetBufferIndex] ||
-        //         mVertexBuffers[targetBufferIndex]->getSizeInBytes() != bufSize)
-        //     {
-        //         reallocateBuffer(targetBufferIndex);
-        //     }
-
-        // Single pass only for now.
-        Ogre::Pass* r2vbPass = mMaterial->getBestTechnique()->getPass(0);
-
-        // Set pass before binding buffers to activate the GPU programs.
-        sceneMgr->_setPass(r2vbPass);
         if (mFirstUpdate)
         {
             bindVerticesOutput(r2vbPass);
             mFirstUpdate = false;
         }
 
-        r2vbPass->_updateAutoParams(sceneMgr->_getAutoParamDataSource(), GPV_GLOBAL);
-
-        // size_t targetBufferIndex = mSourceBufferIndex == 0 ? 0 : 1;
+        if (!mVertexBuffers[mTargetBufferIndex] ||
+            mVertexBuffers[mTargetBufferIndex]->getSizeInBytes() <
+                mVertexData->vertexDeclaration->getVertexSize(0) * mMaxVertexCount)
+        {
+            reallocateBuffer(mTargetBufferIndex);
+        }
 
         // Disable rasterization.
         OGRE_CHECK_GL_ERROR(glEnable(GL_RASTERIZER_DISCARD));
 
-        // Bind shader parameters.
-        RenderSystem* targetRenderSystem = Root::getSingleton().getRenderSystem();
-        if (r2vbPass->hasVertexProgram())
-        {
-            targetRenderSystem->bindGpuProgramParameters(GPT_VERTEX_PROGRAM,
-                                                         r2vbPass->getVertexProgramParameters(), GPV_ALL);
-        }
-        if (r2vbPass->hasFragmentProgram())
-        {
-            targetRenderSystem->bindGpuProgramParameters(GPT_FRAGMENT_PROGRAM,
-                                                         r2vbPass->getFragmentProgramParameters(), GPV_ALL);
-        }
-        if (r2vbPass->hasGeometryProgram())
-        {
-            targetRenderSystem->bindGpuProgramParameters(GPT_GEOMETRY_PROGRAM,
-                                                         r2vbPass->getGeometryProgramParameters(), GPV_ALL);
-        }
-        //TODO add tessellation stages
-
         // Bind source vertex array + target tranform feedback buffer.
-        const GL3PlusHardwareBuffer* targetVertexBuffer = mVertexBuffers[mTargetBufferIndex]->_getImpl<GL3PlusHardwareBuffer>();
+        auto targetVertexBuffer = mVertexBuffers[mTargetBufferIndex]->_getImpl<GL3PlusHardwareBuffer>();
         // OGRE_CHECK_GL_ERROR(glBindBufferBase(GL_TRANSFORM_FEEDBACK_BUFFER, 0, VertexBuffer[mTargetBufferIndex]));
         OGRE_CHECK_GL_ERROR(glBindBufferBase(GL_TRANSFORM_FEEDBACK_BUFFER, 0, targetVertexBuffer->getGLBufferId()));
         // OGRE_CHECK_GL_ERROR(glBindVertexArray(VertexArray[mSourceBufferIndex]));
@@ -233,6 +151,8 @@ namespace Ogre {
             // Use current front buffer to render to back buffer.
             this->getRenderOperation(renderOp);
         }
+
+        RenderSystem* targetRenderSystem = Root::getSingleton().getRenderSystem();
         targetRenderSystem->_render(renderOp);
         // OGRE_CHECK_GL_ERROR(glDrawArrays(GL_POINTS, 0, 1));
 
@@ -258,54 +178,5 @@ namespace Ogre {
 
         // Clear the reset flag.
         mResetRequested = false;
-    }
-
-
-    void GL3PlusRenderToVertexBuffer::reallocateBuffer(size_t index)
-    {
-        assert(index == 0 || index == 1);
-        if (mVertexBuffers[index])
-        {
-            mVertexBuffers[index].reset();
-        }
-
-        // Transform feedback buffer must be at least as large as the
-        // number of output primitives. AMD drivers seem to prefer
-        // that the array be at least one primitive larger than this.
-        mVertexBuffers[index] = HardwareBufferManager::getSingleton().createVertexBuffer(
-            mVertexData->vertexDeclaration->getVertexSize(0), mMaxVertexCount + 1,
-#if OGRE_DEBUG_MODE
-            // Allow reading the contents of the buffer in debug mode.
-            HardwareBuffer::HBU_DYNAMIC
-#else
-            HardwareBuffer::HBU_STATIC_WRITE_ONLY
-#endif
-        );
-    }
-
-
-    String GL3PlusRenderToVertexBuffer::getSemanticVaryingName(VertexElementSemantic semantic, unsigned short index)
-    {
-        switch (semantic)
-        {
-        case VES_POSITION:
-            // Since type of gl_Position cannot be redefined, it is
-            // better to use a custom variable name.
-            // return "gl_Position";
-            return "oPos";
-        case VES_NORMAL:
-            return "oNormal";
-        case VES_TEXTURE_COORDINATES:
-            return String("oUv") + StringConverter::toString(index);
-        case VES_DIFFUSE:
-            return "oColour";
-        case VES_SPECULAR:
-            return "oSecColour";
-            //TODO : Implement more?
-        default:
-            OGRE_EXCEPT(Exception::ERR_RENDERINGAPI_ERROR,
-                        "Unsupported vertex element semantic in render to vertex buffer",
-                        "OgreGL3PlusRenderToVertexBuffer::getSemanticVaryingName");
-        }
     }
 }
