@@ -142,9 +142,7 @@ class _OgreSampleClassExport Sample_Grass : public SdkSample
 
      // build tangent vectors for the ogre head mesh
      MeshPtr headMesh = MeshManager::getSingleton().load("ogrehead.mesh", ResourceGroupManager::DEFAULT_RESOURCE_GROUP_NAME);
-     unsigned short src, dest;
-     if (!headMesh->suggestTangentVectorBuildParams(VES_TANGENT, src, dest))
-         headMesh->buildTangentVectors(VES_TANGENT, src, dest);
+     headMesh->buildTangentVectors();
 
      // put an ogre head in the middle of the field
      Entity* head = mSceneMgr->createEntity("Head", "ogrehead.mesh");
