@@ -116,7 +116,7 @@ void MeshLodTests::blockedWaitForLodGeneration(const MeshPtr& mesh)
     for (int i = 0; i < timeout; i++) 
     {
         OGRE_THREAD_SLEEP(1);
-        wq->processResponses(); // Injects the Lod if ready
+        wq->processMainThreadTasks(); // Injects the Lod if ready
         if (mesh->getNumLodLevels() != 1) {
             success = true;
             break;
