@@ -1916,11 +1916,7 @@ void PlayPen_BuildTangentOnAnimatedMesh::setupContent()
                                                                                HardwareBuffer::HBU_STATIC_WRITE_ONLY,
                                                                                true, true*/); //so we can still read it
     // Build tangent vectors, all our meshes use only 1 texture coordset
-    unsigned short src, dest;
-    if (!pMesh->suggestTangentVectorBuildParams(VES_TANGENT, src, dest))
-    {
-        pMesh->buildTangentVectors(VES_TANGENT, src, dest);
-    }
+    pMesh->buildTangentVectors();
 
     Entity* pEnt = mSceneMgr->createEntity("Ninja", "ninja.mesh");
     pEnt->getAnimationState("Walk")->setEnabled(true);
@@ -5955,11 +5951,7 @@ void PlayPen_StencilShadows::setupContent()
     */
 
     MeshPtr msh = MeshManager::getSingleton().load("knot.mesh", ASSETS_RESOURCE_GROUP);
-    unsigned short src, dest;
-    if (!msh->suggestTangentVectorBuildParams(VES_TANGENT, src, dest))
-    {
-        msh->buildTangentVectors(VES_TANGENT, src, dest);
-    }
+    msh->buildTangentVectors();
     pEnt = mSceneMgr->createEntity( "4", "knot.mesh" );
     pEnt->setMaterialName("Examples/BumpMapping/MultiLightSpecular");
     mTestNode[2] = mSceneMgr->getRootSceneNode()->createChildSceneNode(Vector3(100, 0, 200));
@@ -6387,7 +6379,7 @@ void PlayPen_TextureShadowsIntegrated::setupContent()
     MeshPtr msh = MeshManager::getSingleton().createPlane("Myplane_IS",
                                                           TRANSIENT_RESOURCE_GROUP, plane,
                                                           4500,4500,100,100,true,1,40,40,Vector3::UNIT_Z);
-    msh->buildTangentVectors(VES_TANGENT);
+    msh->buildTangentVectors();
     Entity* pPlaneEnt;
     pPlaneEnt = mSceneMgr->createEntity( "plane", "Myplane_IS" );
     //pPlaneEnt->setMaterialName("Examples/OffsetMapping/Specular");
@@ -6484,7 +6476,7 @@ void PlayPen_TextureShadowsIntegratedPSSM::setupContent()
     MeshPtr msh = MeshManager::getSingleton().createPlane("Myplane",
                                                           TRANSIENT_RESOURCE_GROUP, plane,
                                                           4500,4500,100,100,true,1,40,40,Vector3::UNIT_Z);
-    msh->buildTangentVectors(VES_TANGENT);
+    msh->buildTangentVectors();
     Entity* pPlaneEnt;
     pPlaneEnt = mSceneMgr->createEntity( "plane", "Myplane" );
     pPlaneEnt->setMaterialName("Examples/BumpyMetal");
@@ -6939,9 +6931,7 @@ void PlayPen_PBR::setupContent()
                                                          50, 50, 1, 1, true, 1, 8, 8, Vector3::UNIT_Z);
 
     // build tangent vectors for our mesh
-    unsigned short src, dest;
-    plane->suggestTangentVectorBuildParams(VES_TANGENT, src, dest);
-    plane->buildTangentVectors(VES_TANGENT, src, dest);
+    plane->buildTangentVectors();
 
     auto ground = mSceneMgr->createEntity("Plane");
     ground->setMaterialName("Examples/WoodenFloor");
